refactor(core): decompose runtime hotspots
Split gateway protocols, control routes, model coordination, and llama runtime implementation into focused modules without changing the single-process execution path. Narrow private link dependencies and extend architecture-policy coverage to the extracted implementation files. Refs #121.
