+++ libs/gateway/src/dashboard_routes.cpp
@@ -775,929 +775,13 @@ nlohmann::json build_dashboard_status(const DashboardDeps& deps) {
 
 void register_dashboard_routes(httplib::Server& server, const DashboardDeps& deps,
                                const RouteWrapper& wrap) {
-    server.Post(R"(^/api/inferdeck/v1/optimize/profile$)",
-                wrap([deps](const httplib::Request& req,
-                            httplib::Response& resp) {
-        if (deps.gw.coordinator.active_request_count() > 0 ||
-            deps.gw.coordinator.queued_request_count() > 0) {
-            write_error(resp, 409, "optimization_busy",
-                        "profile analysis waits until active and queued requests are zero");
-            return;
-        }
-        if (deps.gw.swap_tracker && deps.gw.swap_tracker->snapshot().swapping) {
-            write_error(resp, 409, "optimization_busy",
-                        "profile analysis waits until the current model swap finishes");
-            return;
-        }
-        const auto gpu = deps.gpu.latest();
-        if (!gpu.available || gpu.vram_total_mb <= 0.0) {
-            write_error(resp, 503, "optimization_hardware_unavailable",
-                        "GPU VRAM telemetry is required for profile analysis");
-            return;
-        }
-        if (gpu.utilization_pct > 20.0) {
-            write_error(resp, 409, "optimization_busy",
-                        "GPU utilization is above the 20 percent safety threshold");
-            return;
-        }
-        try {
-            const auto body = nlohmann::json::parse(req.body);
-            const std::string model_name = body.value("model", "");
-            auto registered =
-                deps.gw.coordinator.registry().get_info_result(model_name);
-            if (!registered) {
-                write_error(resp, 404, "optimization_model_not_found",
-                            registered.error().message);
-                return;
-            }
-            const auto& info = *registered;
-            if (info.runtime != "llama_cpp") {
-                write_error(resp, 400, "optimization_runtime_unsupported",
-                            "profile analysis currently supports llama.cpp LLM models");
-                return;
-            }
+#include "dashboard_optimize_routes.ipp"
 
-            optimize::ProfileInput input;
-            input.model = model_name;
-            input.total_vram_mb = gpu.vram_total_mb;
-            input.model_file_mb =
-                artifact_size_mb(info.gguf_path) +
-                (info.has_vision ? artifact_size_mb(info.mmproj_path) : 0.0);
-            input.configured_vram_mb = info.vram_required_mb;
-            input.context_per_slot =
-                body.value("contextPerSlot", info.context_size);
-            input.slots = body.value("slots", info.n_slots);
-            input.min_slots = body.value("minSlots", info.min_slots);
-            input.n_batch = body.value("nBatch", 512);
-            input.n_ubatch = body.value("nUbatch", input.n_batch);
-            input.cache_type_k =
-                body.value("cacheTypeK", std::string{"q8_0"});
-            input.cache_type_v =
-                body.value("cacheTypeV", std::string{"q8_0"});
-            input.flash_attention =
-                body.value("flashAttention", std::string{"auto"});
-
-            if (deps.gw.stats_db) {
-                for (const auto& usage : deps.gw.stats_db->model_usage()) {
-                    if (usage.model != model_name ||
-                        usage.total_duration_ms <= 0.0) {
-                        continue;
-                    }
-                    input.observed_tokens_per_second =
-                        static_cast<double>(usage.completion_tokens) /
-                        (usage.total_duration_ms / 1000.0);
-                    break;
-                }
-            }
-
-            const auto result = optimize::recommend_profile(input);
-            nlohmann::json candidates = nlohmann::json::array();
-            for (const auto& candidate : result.candidates) {
-                candidates.push_back(profile_candidate_json(candidate));
-            }
-            write_json(resp, 200, {
-                {"model", model_name},
-                {"mode", "profile_estimate"},
-                {"measured", result.measured},
-                {"observedTokensPerSecond",
-                 input.observed_tokens_per_second},
-                {"modelFileMb", input.model_file_mb},
-                {"totalVramMb", input.total_vram_mb},
-                {"weights", {
-                    {"quality", result.quality_weight},
-                    {"speed", result.speed_weight},
-                    {"parallelism", result.parallelism_weight},
-                    {"headroom", result.headroom_weight},
-                }},
-                {"recommended",
-                 profile_candidate_json(result.recommended)},
-                {"candidates", std::move(candidates)},
-                {"notes", result.notes},
-            });
-        } catch (const std::invalid_argument& error) {
-            write_error(resp, 400, "invalid_optimization_profile",
-                        error.what());
-        } catch (const std::exception& error) {
-            write_error(resp, 422, "optimization_failed", error.what());
-        }
-    }));
-
-    server.Post(R"(^/api/inferdeck/v1/optimize/benchmark$)",
-                wrap([deps](const httplib::Request& req,
-                            httplib::Response& resp) {
-        if (!deps.profile_benchmark) {
-            write_error(resp, 503, "benchmark_unavailable",
-                        "measured profile benchmark is unavailable");
-            return;
-        }
-        const auto gpu = deps.gpu.latest();
-        if (!gpu.available || gpu.vram_total_mb <= 0.0) {
-            write_error(resp, 503, "optimization_hardware_unavailable",
-                        "GPU VRAM telemetry is required for measured benchmarking");
-            return;
-        }
-        if (gpu.utilization_pct > 20.0) {
-            write_error(resp, 409, "optimization_busy",
-                        "GPU utilization is above the 20 percent safety threshold");
-            return;
-        }
-        try {
-            const auto body = nlohmann::json::parse(req.body);
-            const std::string model_name = body.value("model", "");
-            auto registered =
-                deps.gw.coordinator.registry().get_info_result(model_name);
-            if (!registered) {
-                write_error(resp, 404, "optimization_model_not_found",
-                            registered.error().message);
-                return;
-            }
-            const auto& info = *registered;
-            if (info.runtime != "llama_cpp") {
-                write_error(resp, 400, "optimization_runtime_unsupported",
-                            "measured benchmarking currently supports llama.cpp LLM models");
-                return;
-            }
-            optimize::ProfileInput input;
-            input.model = model_name;
-            input.total_vram_mb = gpu.vram_total_mb;
-            input.model_file_mb =
-                artifact_size_mb(info.gguf_path) +
-                (info.has_vision ? artifact_size_mb(info.mmproj_path) : 0.0);
-            input.configured_vram_mb = info.vram_required_mb;
-            input.context_per_slot =
-                body.value("contextPerSlot", info.context_size);
-            input.slots = body.value("slots", info.n_slots);
-            input.min_slots = body.value("minSlots", info.min_slots);
-            input.n_batch = body.value("nBatch", 512);
-            input.n_ubatch = body.value("nUbatch", input.n_batch);
-            input.cache_type_k =
-                body.value("cacheTypeK", std::string{"q8_0"});
-            input.cache_type_v =
-                body.value("cacheTypeV", std::string{"q8_0"});
-            const auto started = deps.profile_benchmark->start(
-                info, input, body.value("candidateLimit", 3));
-            if (!started) {
-                const int status =
-                    started.error().code == foundation::ErrorCode::InvalidArgument
-                        ? 400 : 409;
-                write_error(resp, status, "benchmark_not_started",
-                            started.error().message);
-                return;
-            }
-            write_json(resp, 202,
-                       profile_benchmark_snapshot_json(*started));
-        } catch (const nlohmann::json::exception& error) {
-            write_error(resp, 400, "invalid_benchmark_request", error.what());
-        } catch (const std::exception& error) {
-            write_error(resp, 422, "benchmark_not_started", error.what());
-        }
-    }));
-
-    server.Get(R"(^/api/inferdeck/v1/optimize/benchmark$)",
-               wrap([deps](const httplib::Request&,
-                           httplib::Response& resp) {
-        if (!deps.profile_benchmark) {
-            write_error(resp, 503, "benchmark_unavailable",
-                        "measured profile benchmark is unavailable");
-            return;
-        }
-        write_json(resp, 200, profile_benchmark_snapshot_json(
-            deps.profile_benchmark->snapshot()));
-    }));
-
-    server.Get(R"(^/api/inferdeck/v1/optimize/schedule$)",
-               wrap([deps](const httplib::Request&, httplib::Response& resp) {
-        if (!deps.profile_benchmark_scheduler) {
-            write_error(resp, 503, "optimization_schedule_unavailable",
-                        "scheduled optimization is unavailable");
-            return;
-        }
-        nlohmann::json schedules = nlohmann::json::array();
-        for (const auto& status : deps.profile_benchmark_scheduler->statuses()) {
-            schedules.push_back({
-                {"model", status.model},
-                {"enabled", status.enabled},
-                {"windowStart", status.window_start},
-                {"windowEnd", status.window_end},
-                {"nextRunUnixMs", status.next_run_unix_ms},
-                {"lastStartedUnixMs", status.last_started_unix_ms},
-                {"lastFinishedUnixMs", status.last_finished_unix_ms},
-                {"lastOutcome", status.last_outcome},
-                {"lastMessage", status.last_message},
-            });
-        }
-        write_json(resp, 200, {
-            {"timezone", deps.profile_benchmark_scheduler->timezone_name()},
-            {"schedules", std::move(schedules)},
-        });
-    }));
-
-    server.Post(R"(^/api/inferdeck/v1/optimize/benchmark/cancel$)",
-                wrap([deps](const httplib::Request&,
-                            httplib::Response& resp) {
-        if (!deps.profile_benchmark) {
-            write_error(resp, 503, "benchmark_unavailable",
-                        "measured profile benchmark is unavailable");
-            return;
-        }
-        const auto cancelled = deps.profile_benchmark->cancel();
-        if (!cancelled) {
-            write_error(resp, 409, "benchmark_not_running",
-                        cancelled.error().message);
-            return;
-        }
-        write_json(resp, 202, profile_benchmark_snapshot_json(
-            deps.profile_benchmark->snapshot()));
-    }));
-
-    server.Get(R"(^/api/inferdeck/v1/model-store/search$)", wrap([deps](const httplib::Request& req,
-                                                            httplib::Response& resp) {
-        if (!deps.model_store) {
-            write_error(resp, 503, "model_store_unavailable", "model store is unavailable");
-            return;
-        }
-        const std::string query = req.has_param("q") ? req.get_param_value("q") : "";
-        const std::string runtime = req.has_param("runtime") ? req.get_param_value("runtime") : "";
-        const std::string modality = req.has_param("modality") ? req.get_param_value("modality") : "";
-        int limit = 20;
-        if (req.has_param("limit")) {
-            try {
-                limit = std::clamp(std::stoi(req.get_param_value("limit")), 1, 100);
-            } catch (...) {
-                write_error(resp, 400, "invalid_model_search", "limit must be an integer");
-                return;
-            }
-        }
-        auto result = deps.model_store->search(query, runtime, modality, limit);
-        if (!result) {
-            write_error(resp, result.error().code == foundation::ErrorCode::InvalidArgument ? 400 : 502,
-                        "model_search_failed", result.error().message);
-            return;
-        }
-        write_json(resp, 200, {{"models", std::move(*result)}});
-    }));
-
-    server.Get(R"(^/api/inferdeck/v1/model-store/inspect$)", wrap([deps](const httplib::Request& req,
-                                                             httplib::Response& resp) {
-        if (!deps.model_store) {
-            write_error(resp, 503, "model_store_unavailable", "model store is unavailable");
-            return;
-        }
-        const std::string repo = req.has_param("repo") ? req.get_param_value("repo") : "";
-        auto result = deps.model_store->inspect(repo);
-        if (!result) {
-            const int status = result.error().code == foundation::ErrorCode::InvalidArgument ? 400 :
-                               result.error().code == foundation::ErrorCode::NotFound ? 404 : 502;
-            write_error(resp, status, "model_inspect_failed", result.error().message);
-            return;
-        }
-        write_json(resp, 200, std::move(*result));
-    }));
-
-    server.Get(R"(^/api/inferdeck/v1/model-store/downloads$)", wrap([deps](const httplib::Request&,
-                                                               httplib::Response& resp) {
-        if (!deps.model_store) {
-            write_error(resp, 503, "model_store_unavailable", "model store is unavailable");
-            return;
-        }
-        nlohmann::json downloads = nlohmann::json::array();
-        for (const auto& download : deps.model_store->downloads()) downloads.push_back(to_json(download));
-        write_json(resp, 200, {{"downloads", std::move(downloads)},
-                               {"installed", deps.model_store->installed()},
-                               {"library", deps.model_store->library()}});
-    }));
-
-    server.Post(R"(^/api/inferdeck/v1/model-store/downloads$)", wrap([deps](const httplib::Request& req,
-                                                                httplib::Response& resp) {
-        if (!deps.model_store) {
-            write_error(resp, 503, "model_store_unavailable", "model store is unavailable");
-            return;
-        }
-        try {
-            const auto body = nlohmann::json::parse(req.body);
-            auto result = deps.model_store->install(
-                body.value("repo", ""), body.value("filename", ""),
-                body.value("runtime", ""), body.value("modality", ""),
-                body.value("modelName", ""));
-            if (!result) {
-                const int status = result.error().code == foundation::ErrorCode::AlreadyExists ? 409 : 400;
-                write_error(resp, status, "model_install_failed", result.error().message);
-                return;
-            }
-            write_json(resp, 202, {{"id", *result}, {"state", "queued"}});
-        } catch (const std::exception& error) {
-            write_error(resp, 400, "invalid_model_install", error.what());
-        }
-    }));
-
-    server.Post(R"(^/api/inferdeck/v1/model-store/downloads/([0-9]+)/(cancel|resume)$)",
-                wrap([deps](const httplib::Request& req, httplib::Response& resp) {
-        if (!deps.model_store) {
-            write_error(resp, 503, "model_store_unavailable", "model store is unavailable");
-            return;
-        }
-        const auto id = static_cast<std::uint64_t>(std::stoull(req.matches[1].str()));
-        auto result = req.matches[2].str() == "cancel" ? deps.model_store->cancel(id) :
-                                                        deps.model_store->resume(id);
-        if (!result) {
-            write_error(resp, result.error().code == foundation::ErrorCode::NotFound ? 404 : 409,
-                        "download_control_failed", result.error().message);
-            return;
-        }
-        write_json(resp, 200, {{"ok", true}});
-    }));
-
-    server.Post(R"(^/api/inferdeck/v1/model-store/(remove|archive)$)", wrap([deps](const httplib::Request& req,
-                                                             httplib::Response& resp) {
-        if (!deps.model_store) {
-            write_error(resp, 503, "model_store_unavailable", "model store is unavailable");
-            return;
-        }
-        try {
-            const auto body = nlohmann::json::parse(req.body);
-            auto result = req.matches[1].str() == "archive"
-                ? deps.model_store->archive(body.value("model", ""))
-                : deps.model_store->remove(body.value("model", ""));
-            if (!result) {
-                const int status = result.error().code == foundation::ErrorCode::NotFound ? 404 : 409;
-                write_error(resp, status, "model_retire_failed", result.error().message);
-                return;
-            }
-            write_json(resp, 200, {{"ok", true}});
-        } catch (const std::exception& error) {
-            write_error(resp, 400, "invalid_model_retire", error.what());
-        }
-    }));
-
-    server.Post(R"(^/api/inferdeck/v1/model-store/unregister$)", wrap([deps](const httplib::Request& req,
-                                                                httplib::Response& resp) {
-        if (!deps.config_repository) {
-            write_error(resp, 503, "config_unavailable", "configuration repository is unavailable");
-            return;
-        }
-        try {
-            const auto body = nlohmann::json::parse(req.body);
-            const std::string name = body.value("model", "");
-            if (name.empty()) {
-                write_error(resp, 400, "missing_model", "request body must include model");
-                return;
-            }
-            if (!body.contains("revision") || !body["revision"].is_string()) {
-                write_error(resp, 400, "invalid_model_unregister", "revision is required");
-                return;
-            }
-            const auto info = deps.gw.coordinator.registry().get_info_result(name);
-            if (!info) {
-                write_error(resp, 404, "model_not_found", info.error().message);
-                return;
-            }
-            bool unregistered = false;
-            const auto written = deps.config_repository->transact_active(
-                body["revision"].get<std::string>(),
-                [&](const ConfigSnapshot& snapshot) -> foundation::Result<std::string> {
-                    const auto& original = snapshot.has_active ? snapshot.active : snapshot.base;
-                    const auto root = YAML::Load(original);
-                    if (root["default_model"] &&
-                        root["default_model"].as<std::string>() == name) {
-                        return foundation::Err<std::string>(
-                            foundation::ErrorCode::AlreadyExists,
-                            "change the default model before removing " + name);
-                    }
-                    for (const auto& alias : deps.gw.coordinator.registry().aliases()) {
-                        if (alias.target == name) return foundation::Err<std::string>(
-                            foundation::ErrorCode::AlreadyExists,
-                            "remove or retarget alias " + alias.name + " first");
-                    }
-                    auto updated = remove_model_registry_entry(original, name);
-                    if (!updated) return updated;
-                    const auto removed = deps.gw.coordinator.unregister(name);
-                    if (!removed) return foundation::Err<std::string>(
-                        removed.error().code, removed.error().message);
-                    unregistered = true;
-                    return updated;
-                },
-                [&] {
-                    if (unregistered) deps.gw.coordinator.registry().register_model(*info);
-                });
-            if (!written) {
-                const int status = written.error().code == foundation::ErrorCode::AlreadyExists
-                    ? 409 : written.error().code == foundation::ErrorCode::NotFound ? 404 : 500;
-                write_error(resp, status, "model_unregister_persist_failed",
-                            written.error().message);
-                return;
-            }
-            write_json(resp, 200, {
-                {"ok", true},
-                {"filesDeleted", false},
-                {"restartRequired", false},
-                {"revision", written->revision},
-            });
-        } catch (const std::exception& error) {
-            write_error(resp, 400, "invalid_model_unregister", error.what());
-        }
-    }));
-
-    server.Get(R"(^/api/inferdeck/v1/model-aliases$)", wrap([deps](const httplib::Request&,
-                                                       httplib::Response& resp) {
-        nlohmann::json aliases = nlohmann::json::array();
-        for (const auto& alias : deps.gw.coordinator.registry().aliases()) {
-            aliases.push_back(alias_json(alias));
-        }
-        if (!deps.config_repository) {
-            write_error(resp, 503, "config_unavailable", "configuration repository is unavailable");
-            return;
-        }
-        const auto snapshot = deps.config_repository->snapshot();
-        if (!snapshot) {
-            write_error(resp, 500, "config_read_failed", snapshot.error().message);
-            return;
-        }
-        write_json(resp, 200, {
-            {"aliases", std::move(aliases)},
-            {"revision", snapshot->active_revision},
-        });
-    }));
-
-    server.Put(R"(^/api/inferdeck/v1/model-aliases/([A-Za-z0-9_.-]+)$)",
-               wrap([deps](const httplib::Request& req, httplib::Response& resp) {
-        if (!deps.config_repository) {
-            write_error(resp, 503, "config_unavailable", "configuration repository is unavailable");
-            return;
-        }
-        const std::string name = req.matches[1].str();
-        if (name.empty() || name.size() > 128) {
-            write_error(resp, 400, "invalid_model_alias", "alias name is invalid");
-            return;
-        }
-        try {
-            const auto body = nlohmann::json::parse(req.body);
-            if (!body.contains("revision") || !body["revision"].is_string()) {
-                write_error(resp, 400, "invalid_model_alias", "revision is required");
-                return;
-            }
-            model::ModelAlias requested;
-            requested.name = name;
-            requested.target = body.value("target", "");
-            requested.required_context_size = body.value("requiredContextSize", 0);
-            if (body.contains("requiredCapabilities")) {
-                requested.required_capabilities =
-                    body["requiredCapabilities"].get<std::vector<std::string>>();
-            }
-            std::optional<model::ModelAlias> previous;
-            for (const auto& alias : deps.gw.coordinator.registry().aliases()) {
-                if (alias.name == name) previous = alias;
-            }
-            std::optional<model::ModelAlias> committed;
-            const auto persisted = deps.config_repository->transact_active(
-                body["revision"].get<std::string>(),
-                [&](const ConfigSnapshot& snapshot) -> foundation::Result<std::string> {
-                    const auto applied = deps.gw.coordinator.registry().set_alias(requested);
-                    if (!applied) return foundation::Err<std::string>(
-                        applied.error().code, applied.error().message);
-                    committed = *applied;
-                    return render_aliases(
-                        snapshot.has_active ? snapshot.active : snapshot.base,
-                        deps.gw.coordinator.registry().aliases());
-                },
-                [&] {
-                    if (previous) (void)deps.gw.coordinator.registry().set_alias(*previous);
-                    else (void)deps.gw.coordinator.registry().remove_alias(name);
-                });
-            if (!persisted) {
-                const int status = persisted.error().code == foundation::ErrorCode::AlreadyExists
-                    ? 409 : persisted.error().code == foundation::ErrorCode::InvalidArgument ? 400 : 500;
-                write_error(resp, status,
-                            status == 409 ? "config_conflict" : "model_alias_persist_failed",
-                            persisted.error().message);
-                return;
-            }
-            auto response = alias_json(*committed);
-            response["revision"] = persisted->revision;
-            write_json(resp, previous ? 200 : 201, response);
-        } catch (const std::exception& error) {
-            write_error(resp, 400, "invalid_model_alias", error.what());
-        }
-    }));
-
-    server.Delete(R"(^/api/inferdeck/v1/model-aliases/([A-Za-z0-9_.-]+)$)",
-                  wrap([deps](const httplib::Request& req, httplib::Response& resp) {
-        if (!deps.config_repository) {
-            write_error(resp, 503, "config_unavailable", "configuration repository is unavailable");
-            return;
-        }
-        if (!req.has_header("If-Match") || req.get_header_value("If-Match").empty()) {
-            write_error(resp, 400, "invalid_model_alias", "If-Match revision is required");
-            return;
-        }
-        const std::string name = req.matches[1].str();
-        std::optional<model::ModelAlias> previous;
-        for (const auto& alias : deps.gw.coordinator.registry().aliases()) {
-            if (alias.name == name) previous = alias;
-        }
-        if (!previous) {
-            write_error(resp, 404, "model_alias_not_found", "model alias not found: " + name);
-            return;
-        }
-        const auto persisted = deps.config_repository->transact_active(
-            req.get_header_value("If-Match"),
-            [&](const ConfigSnapshot& snapshot) -> foundation::Result<std::string> {
-                const auto removed = deps.gw.coordinator.registry().remove_alias(name);
-                if (!removed) return foundation::Err<std::string>(
-                    removed.error().code, removed.error().message);
-                return render_aliases(
-                    snapshot.has_active ? snapshot.active : snapshot.base,
-                    deps.gw.coordinator.registry().aliases());
-            },
-            [&] { (void)deps.gw.coordinator.registry().set_alias(*previous); });
-        if (!persisted) {
-            const int status = persisted.error().code == foundation::ErrorCode::AlreadyExists
-                ? 409 : 500;
-            write_error(resp, status,
-                        status == 409 ? "config_conflict" : "model_alias_persist_failed",
-                        persisted.error().message);
-            return;
-        }
-        write_json(resp, 200, {{"ok", true}, {"revision", persisted->revision}});
-    }));
-
-    server.Get(R"(^/api/inferdeck/v1/config$)", wrap([deps](const httplib::Request& req,
-                                               httplib::Response& resp) {
-        (void)req;
-        if (!deps.config_repository) {
-            write_error(resp, 503, "config_unavailable", "configuration path is unavailable");
-            return;
-        }
-        const auto snapshot = deps.config_repository->snapshot();
-        if (!snapshot) {
-            write_error(resp, 500, "config_read_failed", snapshot.error().message);
-            return;
-        }
-        const std::string& desired = snapshot->has_active ? snapshot->active : snapshot->base;
-        write_json(resp, 200, {
-            {"yaml", mask_config_secrets(snapshot->base)},
-            {"revision", snapshot->base_revision},
-            {"activeYaml", mask_config_secrets(desired)},
-            {"activeRevision", snapshot->active_revision},
-            {"runningRevision", deps.running_config_revision},
-            {"hasActiveProfile", snapshot->has_active},
-            {"usingActiveProfile", deps.using_active_config},
-            {"fallbackReason", deps.config_fallback_reason},
-            {"restartRequired", deps.running_config_revision != snapshot->active_revision},
-            {"secretSentinel", "__INFERDECK_SECRET__"},
-        });
-    }));
-
-    server.Put(R"(^/api/inferdeck/v1/config$)", wrap([deps](const httplib::Request& req,
-                                               httplib::Response& resp) {
-        if (maintenance_mode_active(deps.gw)) {
-            write_error(resp, 503, "maintenance_mode",
-                        "configuration cannot change during a measured benchmark");
-            return;
-        }
-        if (!deps.config_repository) {
-            write_error(resp, 503, "config_unavailable", "configuration repository is unavailable");
-            return;
-        }
-        nlohmann::json body;
-        try {
-            body = nlohmann::json::parse(req.body);
-        } catch (const std::exception& error) {
-            write_error(resp, 400, "invalid_json", error.what());
-            return;
-        }
-        if (!body.contains("yaml") || !body["yaml"].is_string() ||
-            !body.contains("revision") || !body["revision"].is_string()) {
-            write_error(resp, 400, "invalid_config_update", "yaml and revision are required");
-            return;
-        }
-        std::string submitted = body["yaml"].get<std::string>();
-        if (submitted.empty() || submitted.size() > 2 * 1024 * 1024) {
-            write_error(resp, 400, "invalid_config_update", "configuration size is invalid");
-            return;
-        }
-        const auto current = deps.config_repository->snapshot();
-        if (!current) {
-            write_error(resp, 500, "config_read_failed", current.error().message);
-            return;
-        }
-        auto written = deps.config_repository->write_base(
-            body["revision"].get<std::string>(), submitted);
-        if (!written) {
-            const int status = written.error().code == foundation::ErrorCode::AlreadyExists
-                ? 409 : written.error().code == foundation::ErrorCode::InvalidArgument ? 400 : 500;
-            write_error(resp, status,
-                        status == 409 ? "config_conflict" :
-                        status == 400 ? "invalid_configuration" : "config_write_failed",
-                        written.error().message);
-            return;
-        }
-        write_json(resp, 200, {
-            {"ok", true},
-            {"revision", written->revision},
-            {"restartRequired", true},
-        });
-    }));
-
-    server.Put(R"(^/api/inferdeck/v1/config/active$)", wrap([deps](const httplib::Request& req,
-                                                      httplib::Response& resp) {
-        if (maintenance_mode_active(deps.gw)) {
-            write_error(resp, 503, "maintenance_mode",
-                        "configuration cannot change during a measured benchmark");
-            return;
-        }
-        if (!deps.config_repository) {
-            write_error(resp, 503, "config_unavailable", "configuration paths are unavailable");
-            return;
-        }
-        nlohmann::json body;
-        try {
-            body = nlohmann::json::parse(req.body);
-        } catch (const std::exception& error) {
-            write_error(resp, 400, "invalid_json", error.what());
-            return;
-        }
-        if (!body.contains("yaml") || !body["yaml"].is_string() ||
-            !body.contains("revision") || !body["revision"].is_string()) {
-            write_error(resp, 400, "invalid_config_update", "yaml and revision are required");
-            return;
-        }
-        std::string submitted = body["yaml"].get<std::string>();
-        if (submitted.empty() || submitted.size() > 2 * 1024 * 1024) {
-            write_error(resp, 400, "invalid_config_update", "configuration size is invalid");
-            return;
-        }
-        const auto current = deps.config_repository->snapshot();
-        if (!current) {
-            write_error(resp, 500, "config_read_failed", current.error().message);
-            return;
-        }
-        auto written = deps.config_repository->write_active(
-            body["revision"].get<std::string>(), submitted);
-        if (!written) {
-            const int status = written.error().code == foundation::ErrorCode::AlreadyExists
-                ? 409 : written.error().code == foundation::ErrorCode::InvalidArgument ? 400 :
-                  written.error().code == foundation::ErrorCode::Unavailable ? 503 : 500;
-            write_error(resp, status,
-                        status == 409 ? "config_conflict" :
-                        status == 400 ? "invalid_configuration" :
-                        status == 503 ? "config_reload_failed" : "config_write_failed",
-                        written.error().message);
-            return;
-        }
-        write_json(resp, 200, {
-            {"ok", true},
-            {"activeRevision", written->revision},
-            {"hasActiveProfile", true},
-            {"restartRequired", false},
-            {"applyScheduled", true},
-        });
-    }));
-
-    server.Delete(R"(^/api/inferdeck/v1/config/active$)", wrap([deps](const httplib::Request& req,
-                                                         httplib::Response& resp) {
-        if (maintenance_mode_active(deps.gw)) {
-            write_error(resp, 503, "maintenance_mode",
-                        "configuration cannot change during a measured benchmark");
-            return;
-        }
-        if (!deps.config_repository) {
-            write_error(resp, 503, "config_unavailable", "configuration repository is unavailable");
-            return;
-        }
-        if (!req.has_header("If-Match") || req.get_header_value("If-Match").empty()) {
-            write_error(resp, 400, "invalid_config_reset", "If-Match active revision is required");
-            return;
-        }
-        auto reset = deps.config_repository->reset_active(req.get_header_value("If-Match"));
-        if (!reset) {
-            const int status = reset.error().code == foundation::ErrorCode::AlreadyExists
-                ? 409 : reset.error().code == foundation::ErrorCode::Unavailable ? 503 : 500;
-            write_error(resp, status,
-                        status == 409 ? "config_conflict" :
-                        status == 503 ? "config_reload_failed" : "config_reset_failed",
-                        reset.error().message);
-            return;
-        }
-        write_json(resp, 200, {
-            {"ok", true},
-            {"removed", reset->changed},
-            {"activeRevision", reset->revision},
-            {"hasActiveProfile", false},
-            {"restartRequired", false},
-            {"applyScheduled", reset->changed},
-        });
-    }));
-
-    server.Get(R"(^/api/inferdeck/v1/status$)", wrap([deps](const httplib::Request& req,
-                                               httplib::Response& resp) {
-        (void)req;
-        resp.set_content(build_dashboard_status(deps).dump(), "application/json");
-    }));
-
-    server.Get(R"(^/api/inferdeck/v1/models$)", wrap([deps](const httplib::Request& req,
-                                               httplib::Response& resp) {
-        (void)req;
-        write_json(resp, 200, build_dashboard_models(deps.gw.coordinator));
-    }));
-
-    server.Get(R"(^/api/inferdeck/v1/usage/daily$)",
-               wrap([deps](const httplib::Request& req,
-                           httplib::Response& resp) {
-        (void)req;
-        write_json(resp, 200, {
-            {"dailyTokenUsage",
-             usage_bucket_json(deps.gw.stats_db->daily_usage(0))},
-            {"dailyTokenUsageAllTime", true},
-        });
-    }));
-
-    server.Get(R"(^/api/inferdeck/v1/jobs$)", wrap([deps](const httplib::Request& req,
-                                             httplib::Response& resp) {
-        int limit = 100;
-        if (req.has_param("limit")) {
-            try { limit = std::clamp(std::stoi(req.get_param_value("limit")), 1, 500); } catch (...) {}
-        }
-        const auto protocol_profile = req.has_param("protocol_profile")
-            ? req.get_param_value("protocol_profile") : std::string{};
-        const auto endpoint = req.has_param("endpoint")
-            ? req.get_param_value("endpoint") : std::string{};
-        resp.set_content(build_dashboard_jobs(
-            *deps.gw.stats_db, limit, protocol_profile, endpoint).dump(),
-            "application/json");
-    }));
-
-    server.Post(R"(^/api/inferdeck/v1/models/load$)", wrap([deps](const httplib::Request& req,
-                                                     httplib::Response& resp) {
-        auto body = req.body.empty() ? nlohmann::json::object() : nlohmann::json::parse(req.body);
-        const std::string model_name = body.value("model", body.value("name", ""));
-        if (model_name.empty()) {
-            write_error(resp, 400, "missing_model", "request body must include model");
-            return;
-        }
-        auto started = start_swap_async(deps.gw, model_name);
-        write_json(resp, started.status, started.body);
-    }));
-
-    server.Post(R"(^/api/inferdeck/v1/models/unload$)", wrap([deps](const httplib::Request& req,
-                                                       httplib::Response& resp) {
-        const auto body = req.body.empty() ? nlohmann::json::object() : nlohmann::json::parse(req.body);
-        const auto current = deps.gw.coordinator.get_loaded_model();
-        const std::string requested_model = body.value("model", current.value_or(""));
-        const auto resolved = requested_model.empty()
-            ? foundation::Ok(std::string{})
-            : deps.gw.coordinator.registry().resolve(requested_model);
-        if (!resolved) {
-            write_error(resp, 404, "model_not_found", resolved.error().message);
-            return;
-        }
-        const std::string& model_name = *resolved;
-        if (!model_name.empty() && maintenance_blocks_model(deps.gw, model_name)) {
-            write_error(resp, 503, "maintenance_mode",
-                        "measured model optimization is using the same compute resource");
-            return;
-        }
-        auto result = model_name.empty() ? foundation::Ok() : deps.gw.coordinator.unload(model_name);
-        if (!result) {
-            write_error(resp, 500, "unload_failed", result.error().message);
-            return;
-        }
-        if (deps.gw.events && !model_name.empty()) {
-            deps.gw.events->publish("model", nlohmann::json{
-                {"state", "unloaded"},
-                {"from", model_name},
-                {"to", ""},
-                {"durationMs", 0.0},
-                {"error", ""},
-            }.dump());
-        }
-        write_json(resp, 200, {{"ok", true}, {"status", "stopped"},
-                               {"model", requested_model},
-                               {"resolvedModel", model_name}});
-    }));
-
-    server.Get(R"(^/api/inferdeck/v1/pricing$)", wrap([deps](const httplib::Request& req,
-                                                httplib::Response& resp) {
-        (void)req;
-        nlohmann::json pricing = nlohmann::json::array();
-        std::ifstream file(deps.pricing_file);
-        if (file.is_open()) {
-            try {
-                pricing = nlohmann::json::parse(file);
-                if (!pricing.is_array()) pricing = nlohmann::json::array();
-            } catch (const std::exception& error) {
-                LOG_WARN("pricing_file_invalid", "path={} error={}", deps.pricing_file, error.what());
-                pricing = nlohmann::json::array();
-            }
-        } else {
-            LOG_WARN("pricing_file_unreadable", "path={} file_defaults_unavailable=true", deps.pricing_file);
-        }
-        for (const auto& name : deps.gw.coordinator.registry().list()) {
-            const auto info = deps.gw.coordinator.registry().get_info_result(name);
-            if (!info || (!info->prompt_price_per_million &&
-                          !info->cached_prompt_price_per_million &&
-                          !info->completion_price_per_million)) continue;
-            auto existing = std::find_if(pricing.begin(), pricing.end(), [&](const nlohmann::json& entry) {
-                return entry.value("model_name", "") == name;
-            });
-            nlohmann::json value = existing == pricing.end()
-                ? nlohmann::json{
-                    {"model_name", name},
-                    {"currency", "USD"},
-                    {"prompt_price_per_million",
-                     info->prompt_price_per_million.value_or(0.0)},
-                    {"cached_prompt_price_per_million",
-                     info->cached_prompt_price_per_million.value_or(
-                         info->prompt_price_per_million.value_or(0.0))},
-                    {"completion_price_per_million",
-                     info->completion_price_per_million.value_or(0.0)}}
-                : *existing;
-            if (info->prompt_price_per_million) {
-                value["prompt_price_per_million"] =
-                    *info->prompt_price_per_million;
-            }
-            if (info->cached_prompt_price_per_million) {
-                value["cached_prompt_price_per_million"] =
-                    *info->cached_prompt_price_per_million;
-            }
-            if (info->completion_price_per_million) {
-                value["completion_price_per_million"] =
-                    *info->completion_price_per_million;
-            }
-            value["source"] = "model_settings";
-            if (existing == pricing.end()) pricing.push_back(std::move(value));
-            else *existing = std::move(value);
-        }
-        for (const auto& alias : deps.gw.coordinator.registry().aliases()) {
-            auto target = std::find_if(pricing.begin(), pricing.end(), [&](const nlohmann::json& entry) {
-                return entry.value("model_name", "") == alias.target;
-            });
-            if (target == pricing.end()) continue;
-            nlohmann::json value = *target;
-            value["model_name"] = alias.name;
-            value["source"] = "model_alias";
-            auto existing = std::find_if(pricing.begin(), pricing.end(), [&](const nlohmann::json& entry) {
-                return entry.value("model_name", "") == alias.name;
-            });
-            if (existing == pricing.end()) pricing.push_back(std::move(value));
-            else *existing = std::move(value);
-        }
-        resp.set_content(pricing.dump(), "application/json");
-    }));
-
-    server.Get(R"(^/api/inferdeck/v1/logs$)", wrap([deps](const httplib::Request& req,
-                                             httplib::Response& resp) {
-        std::size_t limit = 250;
-        if (req.has_param("limit")) {
-            try { limit = std::clamp<std::size_t>(std::stoul(req.get_param_value("limit")), 1, 1000); } catch (...) {}
-        }
-        const auto lines = bounded_log_tail(
-            deps.log_file.empty() ? "logs/gateway.log" : deps.log_file,
-            limit);
-        nlohmann::json logs = nlohmann::json::array();
-        for (const auto& item : lines) {
-            logs.push_back({{"message", item}});
-        }
-        resp.set_content(nlohmann::json{{"logs", logs}}.dump(), "application/json");
-    }));
-
-    server.Get(R"(^/api/inferdeck/v1/events/stream$)", wrap([deps](const httplib::Request& req,
-                                                 httplib::Response& resp) {
-        (void)req;
-        if (!deps.gw.events) {
-            resp.status = 503;
-            return;
-        }
-        auto lease = std::make_shared<DashboardStreamLease>();
-        if (!lease->acquire()) {
-            write_error(resp, 503, "too_many_event_streams",
-                        "dashboard event stream limit reached");
-            return;
-        }
-        auto sub = deps.gw.events->subscribe(64);
-        LOG_INFO("sse_subscribe", "subscribers={}", deps.gw.events->subscriber_count());
-        resp.set_header("Cache-Control", "no-cache");
-        resp.set_header("X-Accel-Buffering", "no");
-        resp.set_chunked_content_provider(
-            "text/event-stream",
-            [sub, lease](std::size_t, httplib::DataSink& sink) -> bool {
-                if (sub->closed()) return false;
-                auto ev = sub->wait_for(std::chrono::milliseconds{2000});
-                if (!sink.is_writable()) return false;
-                std::string out;
-                if (!ev) {
-                    out = ": \n\n";
-                } else {
-                    out = "event: " + ev->name + "\ndata: " + ev->data + "\n\n";
-                }
-                return sink.write(out.data(), out.size());
-            },
-            [sub, lease, deps](bool) {
-                sub->close();
-                lease->release();
-                LOG_INFO("sse_unsubscribe", "subscribers={}",
-                         deps.gw.events ? deps.gw.events->subscriber_count() : 0);
-            });
-    }));
+#include "dashboard_model_store_routes.ipp"
+#include "dashboard_alias_routes.ipp"
+#include "dashboard_config_routes.ipp"
+#include "dashboard_status_routes.ipp"
+#include "dashboard_observability_routes.ipp"
 }
 
 } // namespace inferdeck::gateway
