+++ libs/gateway/src/media_routes.cpp
@@ -502,771 +502,10 @@ foundation::Result<void> cancel_media_job(std::uint64_t id) {
     return foundation::Ok();
 }
 
-void handle_image_generations(const httplib::Request& req, httplib::Response& resp,
-                              const GatewayDeps& deps) {
-    const auto observation = observe_request(
-        req, resp, deps, "image_generation", false);
-    if (!require_json_media_type(req, resp)) return;
-    nlohmann::json body;
-    try { body = nlohmann::json::parse(req.body); }
-    catch (const std::exception& error) { write_error(resp, 400, "invalid_json", error.what()); return; }
-    const bool derivative =
-        deps.compatibility_profile == CompatibilityProfile::OpenAIDerivative;
-    if (!derivative) {
-        static constexpr std::array<std::string_view, 14> supported_fields{
-            "model", "prompt", "background", "moderation", "n",
-            "output_compression", "output_format", "partial_images",
-            "quality", "response_format", "size", "stream", "style", "user",
-        };
-        if (!body.is_object()) {
-            write_error(resp, 400, "invalid_request_error",
-                        "request body must be an object");
-            return;
-        }
-        for (const auto& field : body.items()) {
-            if (std::find(supported_fields.begin(), supported_fields.end(),
-                          field.key()) == supported_fields.end()) {
-                write_error(resp, 400, "unsupported_parameter",
-                            "unsupported Images parameter: " +
-                                field.key(),
-                            field.key());
-                return;
-            }
-        }
-        const auto compatible_default =
-            [&body, &resp](std::string_view field, std::string_view value) {
-            if (!body.contains(field) || body[field].is_null()) return true;
-            if (!body[field].is_string() ||
-                body[field].get<std::string>() != value) {
-                write_error(resp, 400, "unsupported_parameter",
-                            std::string(field) + " is not supported by the native image runtime",
-                            std::string(field));
-                return false;
-            }
-            return true;
-        };
-        if (!compatible_default("background", "auto") ||
-            !compatible_default("moderation", "auto") ||
-            !compatible_default("output_format", "png") ||
-            !compatible_default("quality", "auto") ||
-            !compatible_default("response_format", "b64_json")) {
-            return;
-        }
-        if (body.contains("style") && !body["style"].is_null()) {
-            write_error(resp, 400, "unsupported_parameter",
-                        "style is not supported by the native image runtime",
-                        "style");
-            return;
-        }
-        if (body.contains("output_compression") &&
-            !body["output_compression"].is_null()) {
-            write_error(resp, 400, "unsupported_parameter",
-                        "output_compression is not supported for PNG output",
-                        "output_compression");
-            return;
-        }
-        if (body.contains("stream") && !body["stream"].is_null() &&
-            (!body["stream"].is_boolean() || body["stream"].get<bool>())) {
-            write_error(resp, 400, "unsupported_parameter",
-                        "streaming image generation is not supported",
-                        "stream");
-            return;
-        }
-        if (body.contains("partial_images") &&
-            !body["partial_images"].is_null() &&
-            (!body["partial_images"].is_number_integer() ||
-             body["partial_images"].get<int>() != 0)) {
-            write_error(resp, 400, "unsupported_parameter",
-                        "partial_images requires unsupported image streaming",
-                        "partial_images");
-            return;
-        }
-        if (body.contains("user") &&
-            (!body["user"].is_string() ||
-             body["user"].get_ref<const std::string&>().size() > 64)) {
-            write_error(resp, 400, "invalid_user",
-                        "user must be a string of at most 64 characters",
-                        "user");
-            return;
-        }
-    }
-    if (!body.is_object()) {
-        write_error(resp, 400, "invalid_image_request",
-                    "request body must be an object");
-        return;
-    }
-    if (!body.contains("prompt") || !body["prompt"].is_string()) {
-        write_error(resp, 400, "invalid_image_request",
-                    "prompt must be a string", "prompt");
-        return;
-    }
-    if (body.contains("model") && !body["model"].is_null() &&
-        !body["model"].is_string()) {
-        write_error(resp, 400, "invalid_image_request",
-                    "model must be a string", "model");
-        return;
-    }
-    if (body.contains("n") && !body["n"].is_null() &&
-        !body["n"].is_number_integer()) {
-        write_error(resp, 400, "invalid_image_request",
-                    "n must be an integer", "n");
-        return;
-    }
-    if (body.contains("size") && !body["size"].is_null() &&
-        !body["size"].is_string()) {
-        write_error(resp, 400, "invalid_image_request",
-                    "size must be a string", "size");
-        return;
-    }
-    const std::string model_name =
-        body.contains("model") && !body["model"].is_null()
-            ? body["model"].get<std::string>() : deps.default_model;
-    model::ImageGenerationRequest request;
-    request.prompt = body["prompt"].get<std::string>();
-    request.negative_prompt = body.value("negative_prompt", "");
-    request.count = body.contains("n") && !body["n"].is_null()
-        ? body["n"].get<int>() : 1;
-    request.seed = body.value("seed", std::int64_t{-1});
-    request.steps = body.value("steps", 20);
-    request.guidance_scale = body.value("guidance_scale", 7.0f);
-    const std::string size =
-        body.contains("size") && !body["size"].is_null()
-            ? body["size"].get<std::string>() : "1024x1024";
-    char trailing = '\0';
-    if (std::sscanf(size.c_str(), "%dx%d%c", &request.width, &request.height, &trailing) != 2 ||
-        model_name.empty() || request.prompt.empty() || request.prompt.size() > 32000 ||
-        request.negative_prompt.size() > 32768 || request.count < 1 || request.count > 4 ||
-        request.width < 256 || request.height < 256 || request.width > 2048 || request.height > 2048 ||
-        request.width % 64 != 0 || request.height % 64 != 0 || request.steps < 1 || request.steps > 200 ||
-        !std::isfinite(request.guidance_scale) || request.guidance_scale < 0.0f || request.guidance_scale > 50.0f) {
-        write_error(resp, 400, "invalid_image_request", "invalid model, prompt, size, count, seed, steps, or guidance scale");
-        return;
-    }
-    const auto resolved_model = resolve_model_name(deps, model_name);
-    if (!resolved_model) {
-        write_error(resp, 404, "model_not_found", resolved_model.error().message);
-        return;
-    }
-    auto job = begin_job(model_name, "image");
-    if (derivative) {
-        resp.set_header("X-InferDeck-Job-Id", std::to_string(job->id));
-    }
-    const std::string& runtime_model = resolved_model->resolved;
-    auto slot = acquire_media_slot(req, deps, runtime_model, job);
-    if (!slot) {
-        const int status = status_for(slot.error().code);
-        const int internal_status = internal_status_for(slot.error().code);
-        write_error(resp, status, "image_admission_failed", slot.error().message);
-        record_media(deps, model_name, 0, internal_status, -1,
-                     0.0, 0, observation);
-        finish_job(job, internal_status == 499 ? "cancelled" : "failed");
-        return;
-    }
-    SlotGuard guard{&deps.coordinator, runtime_model, *slot};
-    auto result = deps.coordinator.generate_images(runtime_model, *slot, request,
-        [&req, &deps, &model_name, job](int progress) {
-            update_job(job, progress);
-            if (deps.events) deps.events->publish("progress", nlohmann::json{{"id", job->id}, {"model", model_name}, {"modality", "image"}, {"progress", progress}}.dump());
-            return !req.is_connection_closed() && !job->cancelled->load();
-        });
-    if (!result) {
-        const bool cancelled = job->cancelled->load() ||
-            result.error().code == foundation::ErrorCode::Cancelled;
-        const int status = cancelled ? 408 : status_for(result.error().code);
-        const int internal_status = cancelled ? 499 : status;
-        write_error(resp, status, "image_generation_failed", result.error().message);
-        record_media(deps, model_name, 0, internal_status, *slot,
-                     0.0, 0, observation);
-        finish_job(job, cancelled ? "cancelled" : "failed");
-        return;
-    }
-    nlohmann::json data = nlohmann::json::array();
-    for (const auto& image : result->png_images) data.push_back({{"b64_json", base64(image)}});
-    write_json(resp, 200, {{"created", std::chrono::duration_cast<std::chrono::seconds>(
-                                std::chrono::system_clock::now().time_since_epoch()).count()},
-                           {"output_format", "png"},
-                           {"data", std::move(data)}});
-    auto completed_observation = observation;
-    completed_observation.output_image_count =
-        static_cast<int>(result->png_images.size());
-    record_media(deps, model_name, result->duration_ms, 200, *slot,
-                 0.0, 0, completed_observation);
-    finish_job(job, "completed");
-}
+#include "image_routes.ipp"
 
-void handle_audio_speech(const httplib::Request& req, httplib::Response& resp,
-                         const GatewayDeps& deps) {
-    if (!require_json_media_type(req, resp)) return;
-    nlohmann::json body;
-    try { body = nlohmann::json::parse(req.body); }
-    catch (const std::exception& error) { write_error(resp, 400, "invalid_json", error.what()); return; }
-    const bool derivative =
-        deps.compatibility_profile == CompatibilityProfile::OpenAIDerivative;
-    if (!derivative && body.is_object()) {
-        static const std::unordered_set<std::string> fields{
-            "model", "input", "voice", "instructions", "response_format",
-            "speed", "stream_format",
-        };
-        for (const auto& field : body.items()) {
-            if (!fields.contains(field.key())) {
-                write_error(resp, 400, "unsupported_parameter",
-                            "unsupported speech parameter: " + field.key(),
-                            field.key());
-                return;
-            }
-        }
-    }
-    if (!body.is_object()) {
-        write_error(resp, 400, "invalid_speech_request",
-                    "request body must be an object");
-        return;
-    }
-    if (!body.contains("model") || !body["model"].is_string() ||
-        body["model"].get_ref<const std::string&>().empty()) {
-        write_error(resp, 400, "invalid_speech_request",
-                    "model must be a non-empty string", "model");
-        return;
-    }
-    if (!body.contains("input") || !body["input"].is_string() ||
-        body["input"].get_ref<const std::string&>().empty()) {
-        write_error(resp, 400, "invalid_speech_request",
-                    "input must be a non-empty string", "input");
-        return;
-    }
-    if (!body.contains("voice")) {
-        write_error(resp, 400, "invalid_speech_request",
-                    "voice is required", "voice");
-        return;
-    }
-    const std::string model_name = body["model"].get<std::string>();
-    model::SpeechRequest request;
-    request.input = body["input"].get<std::string>();
-    if (body["voice"].is_string()) {
-        request.voice = body["voice"].get<std::string>();
-    } else if (body["voice"].is_object() && body["voice"].size() == 1 &&
-               body["voice"].contains("id") &&
-               body["voice"]["id"].is_string()) {
-        request.voice = body["voice"]["id"].get<std::string>();
-    }
-    if (request.voice.empty()) {
-        write_error(resp, 400, "invalid_speech_request",
-                    "voice must be a non-empty string or voice ID object",
-                    "voice");
-        return;
-    }
-    if (body.contains("speed")) {
-        if (!body["speed"].is_number()) {
-            write_error(resp, 400, "invalid_speech_request",
-                        "speed must be a number between 0.25 and 4.0",
-                        "speed");
-            return;
-        }
-        const double speed = body["speed"].get<double>();
-        if (!std::isfinite(speed) || speed < 0.25 || speed > 4.0) {
-            write_error(resp, 400, "invalid_speech_request",
-                        "speed must be a number between 0.25 and 4.0",
-                        "speed");
-            return;
-        }
-        request.speed = static_cast<float>(speed);
-    }
-    if (body.contains("instructions")) {
-        if (!body["instructions"].is_string()) {
-            write_error(resp, 400, "invalid_speech_request",
-                        "instructions must be a string", "instructions");
-            return;
-        }
-        const auto instructions = body["instructions"].get<std::string>();
-        if (utf8_character_count(instructions) > 4096) {
-            write_error(resp, 400, "invalid_speech_request",
-                        "instructions must contain at most 4096 characters",
-                        "instructions");
-            return;
-        }
-        if (!instructions.empty()) {
-            write_error(resp, 400, "unsupported_parameter",
-                        "instructions are not supported by this native speech runtime",
-                        "instructions");
-            return;
-        }
-    }
-    if (utf8_character_count(request.input) > 4096) {
-        write_error(resp, 400, "invalid_speech_request",
-                    "input must contain at most 4096 characters", "input");
-        return;
-    }
-    if (body.contains("response_format") && !body["response_format"].is_string()) {
-        write_error(resp, 400, "invalid_speech_request",
-                    "response_format must be a string", "response_format");
-        return;
-    }
-    request.format = body.value("response_format", std::string{"mp3"});
-    static const std::array formats{"mp3", "opus", "aac", "flac", "wav", "pcm"};
-    if (std::find(formats.begin(), formats.end(), request.format) == formats.end()) {
-        write_error(resp, 400, "unsupported_response_format",
-                    "response_format must be mp3, opus, aac, flac, wav, or pcm",
-                    "response_format");
-        return;
-    }
-    if (body.contains("stream_format") && !body["stream_format"].is_string()) {
-        write_error(resp, 400, "invalid_speech_request",
-                    "stream_format must be a string", "stream_format");
-        return;
-    }
-    const std::string stream_format = body.value("stream_format", std::string{"audio"});
-    if (stream_format != "audio" && stream_format != "sse") {
-        write_error(resp, 400, "unsupported_stream_format",
-                    "stream_format must be audio or sse", "stream_format");
-        return;
-    }
-    if (stream_format == "sse") {
-        write_error(resp, 400, "unsupported_stream_format",
-                    "SSE speech streaming is not available for this native runtime",
-                    "stream_format");
-        return;
-    }
-    const auto resolved_model = resolve_model_name(deps, model_name);
-    if (!resolved_model) {
-        write_error(resp, 404, "model_not_found", resolved_model.error().message);
-        return;
-    }
-    const auto info = deps.coordinator.registry().get_info_result(resolved_model->resolved);
-    if (!info || !info->supports("audio_speech")) {
-        write_error(resp, 400, "unsupported_capability",
-                    "model does not support audio speech generation");
-        return;
-    }
-    const bool wav_runtime =
-        info && (info->runtime == "sherpa_onnx" ||
-                 info->runtime == "windows_sapi");
-    const auto input_characters = utf8_character_count(request.input);
-    const auto observation = observe_request(
-        req, resp, deps, "audio_speech", request.format != "wav");
-    if (wav_runtime && request.format != "wav" && request.format != "pcm") {
-        write_error(resp, 400, "unsupported_response_format",
-                    "this native speech runtime supports wav and pcm responses");
-        return;
-    }
-    const std::string& runtime_model = resolved_model->resolved;
-    const auto validation = deps.coordinator.validate_speech_request(
-        runtime_model, request);
-    if (!validation) {
-        const int status = status_for(validation.error().code);
-        write_error(resp, status,
-                    status == 400 ? "invalid_speech_request"
-                                  : "speech_validation_failed",
-                    validation.error().message);
-        return;
-    }
-    const auto validation_deadline = std::chrono::steady_clock::now() +
-        std::chrono::seconds{30};
-    const std::function<bool()> cancelled = [&req] {
-        return req.is_connection_closed();
-    };
-    const auto prepared = ensure_model_loaded(
-        deps, runtime_model, validation_deadline, cancelled);
-    if (!prepared.ok) {
-        write_error(resp, prepared.status, "speech_admission_failed",
-                    prepared.message);
-        return;
-    }
-    VoiceSessionGuard voice_session(req, deps);
-    auto job = begin_job(model_name, "audio_speech");
-    if (deps.compatibility_profile == CompatibilityProfile::OpenAIDerivative) {
-        resp.set_header("X-InferDeck-Job-Id", std::to_string(job->id));
-    }
-    auto slot = acquire_media_slot(req, deps, runtime_model, job);
-    if (!slot) {
-        const int status = status_for(slot.error().code);
-        const int internal_status = internal_status_for(slot.error().code);
-        write_error(resp, status, "speech_admission_failed", slot.error().message);
-        record_media(deps, model_name, 0, internal_status, -1,
-                     0.0, 0, observation);
-        finish_job(job, internal_status == 499 ? "cancelled" : "failed");
-        return;
-    }
-    SlotGuard guard{&deps.coordinator, runtime_model, *slot};
-    if (request.format == "wav") {
-        try {
-            auto result = deps.coordinator.synthesize(
-                runtime_model, *slot, request,
-                [&req, job](const std::byte*, std::size_t) {
-                    return !req.is_connection_closed() &&
-                           !job->cancelled->load();
-                });
-            if (!result) {
-                const bool was_cancelled = job->cancelled->load() ||
-                    result.error().code == foundation::ErrorCode::Cancelled;
-                const int status = was_cancelled
-                    ? 408
-                    : status_for(result.error().code);
-                write_error(resp, status, "speech_generation_failed",
-                            result.error().message);
-                record_media(deps, model_name, 0,
-                             was_cancelled ? 499 : status, *slot,
-                             0.0, 0, observation);
-                finish_job(job, was_cancelled ? "cancelled" : "failed");
-                return;
-            }
-            if (result->bytes.empty()) {
-                write_error(resp, 500, "speech_generation_failed",
-                            "speech runtime returned no audio");
-                record_media(deps, model_name, result->duration_ms, 500, *slot,
-                             0.0, 0, observation);
-                finish_job(job, "failed");
-                return;
-            }
-            resp.set_content(
-                std::string(
-                    reinterpret_cast<const char*>(result->bytes.data()),
-                    result->bytes.size()),
-                result->content_type.empty() ? "audio/wav"
-                                             : result->content_type);
-            auto completed_observation = observation;
-            completed_observation.output_audio_seconds =
-                result->output_audio_seconds;
-            record_media(deps, model_name, result->duration_ms, 200, *slot,
-                         0.0, input_characters, completed_observation);
-            finish_job(job, "completed");
-            return;
-        } catch (const std::exception& error) {
-            write_error(resp, 500, "speech_generation_failed", error.what());
-            record_media(deps, model_name, 0, 500, *slot,
-                         0.0, 0, observation);
-            finish_job(job, "failed");
-            return;
-        }
-    }
-    auto state = std::make_shared<SpeechStreamState>(deps);
-    state->coordinator = &deps.coordinator;
-    state->model = runtime_model;
-    state->requested_model = model_name;
-    state->slot = *slot;
-    state->job = job;
-    state->input_characters = input_characters;
-    state->observation = observation;
-    state->session_key = voice_session.key();
-    state->session_token = voice_session.token();
-    try {
-        state->worker = std::thread([state, request] {
-            try {
-                auto result = state->coordinator->synthesize(
-                    state->model, state->slot, request,
-                    [state](const std::byte* data, std::size_t size) {
-                        try {
-                            if (state->aborted.load() || state->job->cancelled->load()) return false;
-                            if (size > 0) {
-                                if (!data) {
-                                    state->failed.store(true);
-                                    return false;
-                                }
-                                std::unique_lock lock(state->mutex);
-                                auto has_capacity = [state, size] {
-                                    const bool byte_capacity =
-                                        state->chunks.empty() ||
-                                        (size <= state->max_pending_bytes &&
-                                         state->pending_bytes <=
-                                             state->max_pending_bytes - size);
-                                    return state->chunks.size() <
-                                               state->max_pending_chunks &&
-                                           byte_capacity;
-                                };
-                                while (!state->aborted.load() &&
-                                       !state->job->cancelled->load() &&
-                                       !has_capacity()) {
-                                    state->cv.wait_for(
-                                        lock, std::chrono::milliseconds{100});
-                                }
-                                if (state->aborted.load() ||
-                                    state->job->cancelled->load()) {
-                                    return false;
-                                }
-                                state->chunks.emplace_back(reinterpret_cast<const char*>(data), size);
-                                state->pending_bytes += size;
-                                state->streamed_bytes = true;
-                                state->cv.notify_one();
-                            }
-                            return !state->aborted.load() && !state->job->cancelled->load();
-                        } catch (...) {
-                            state->failed.store(true);
-                            return false;
-                        }
-                    });
-                if (!result) {
-                    state->failed.store(!state->job->cancelled->load());
-                    if (state->job->cancelled->load()) state->aborted.store(true);
-                } else {
-                    state->duration_ms = result->duration_ms;
-                    state->output_audio_seconds = result->output_audio_seconds;
-                    std::lock_guard lock(state->mutex);
-                    if (!state->streamed_bytes && !result->bytes.empty()) {
-                        state->chunks.emplace_back(
-                            reinterpret_cast<const char*>(result->bytes.data()),
-                            result->bytes.size());
-                        state->pending_bytes += result->bytes.size();
-                    }
-                }
-            } catch (...) {
-                state->failed.store(true);
-            }
-            state->finished.store(true);
-            state->cv.notify_all();
-        });
-    } catch (const std::exception& error) {
-        write_error(resp, 500, "speech_generation_failed", error.what());
-        record_media(deps, model_name, 0, 500, *slot,
-                     0.0, 0, observation);
-        finish_job(job, "failed");
-        return;
-    }
-    guard.disarm();
-    voice_session.keep();
-    const std::string content_type = request.format == "wav" ? "audio/wav" :
-                                     request.format == "pcm" ? "audio/pcm" :
-                                     request.format == "opus" ? "audio/ogg" :
-                                     request.format == "aac" ? "audio/aac" :
-                                     request.format == "flac" ? "audio/flac" : "audio/mpeg";
-    resp.set_chunked_content_provider(
-        content_type,
-        [state](std::size_t, httplib::DataSink& sink) {
-            std::unique_lock lock(state->mutex);
-            state->cv.wait_for(lock, std::chrono::seconds(2), [state] {
-                return !state->chunks.empty() || state->finished.load() || state->aborted.load();
-            });
-            if (!state->chunks.empty()) {
-                std::string chunk = std::move(state->chunks.front());
-                state->pending_bytes -= chunk.size();
-                state->chunks.pop_front();
-                lock.unlock();
-                state->cv.notify_all();
-                if (!sink.write(chunk.data(), chunk.size())) {
-                    state->finish(499);
-                    return false;
-                }
-                return true;
-            }
-            if (state->finished.load()) {
-                const bool failed = state->failed.load();
-                const bool aborted = state->aborted.load();
-                lock.unlock();
-                state->finish(aborted ? 499 : failed ? 500 : 200);
-                sink.done();
-                return false;
-            }
-            return true;
-        },
-        [state](bool success) {
-            state->finish(!success || state->aborted.load()
-                              ? 499
-                              : state->failed.load() ? 500 : 200);
-        });
-}
+#include "speech_routes.ipp"
 
-void handle_audio_transcriptions(const httplib::Request& req, httplib::Response& resp,
-                                 const GatewayDeps& deps) {
-    const auto observation = observe_request(
-        req, resp, deps, "audio_transcription", false);
-    if (!req.is_multipart_form_data() || !req.form.has_file("file") || !req.form.has_field("model")) {
-        write_error(resp, 400, "invalid_transcription_request",
-                    "multipart file and model fields are required",
-                    !req.form.has_file("file") ? "file" : "model");
-        return;
-    }
-    const auto& file = req.form.files.find("file")->second;
-    if (file.content.empty() || file.content.size() > 25 * 1024 * 1024) {
-        write_error(resp, 400, "invalid_audio",
-                    "audio must be between 1 byte and 25 MB", "file");
-        return;
-    }
-    if (req.form.get_file_count("file") != 1 ||
-        req.form.get_field_count("model") != 1) {
-        write_error(resp, 400, "invalid_transcription_request",
-                    "file and model must each be supplied exactly once",
-                    req.form.get_file_count("file") != 1 ? "file" : "model");
-        return;
-    }
-    static const std::unordered_set<std::string> strict_fields{
-        "model", "chunking_strategy", "include", "include[]", "keywords",
-        "keywords[]", "known_speaker_names", "known_speaker_names[]",
-        "known_speaker_references", "known_speaker_references[]", "language",
-        "languages", "languages[]", "prompt", "response_format", "stream",
-        "temperature", "timestamp_granularities", "timestamp_granularities[]",
-    };
-    if (deps.compatibility_profile != CompatibilityProfile::OpenAIDerivative) {
-        for (const auto& [name, _] : req.form.fields) {
-            if (!strict_fields.contains(name)) {
-                write_error(resp, 400, "unsupported_parameter",
-                            "unsupported transcription parameter: " + name,
-                            name);
-                return;
-            }
-        }
-        for (const auto& [name, _] : req.form.files) {
-            if (name != "file") {
-                write_error(resp, 400, "unsupported_parameter",
-                            "unsupported transcription file parameter: " + name,
-                            name);
-                return;
-            }
-        }
-    }
-    const std::string model_name = req.form.get_field("model");
-    if (model_name.empty()) {
-        write_error(resp, 400, "invalid_transcription_request",
-                    "model must not be empty", "model");
-        return;
-    }
-    const auto single_field = [&req, &resp](const std::string& name) {
-        if (req.form.get_field_count(name) > 1) {
-            write_error(resp, 400, "invalid_transcription_request",
-                        name + " must be supplied at most once", name);
-            return false;
-        }
-        return true;
-    };
-    for (const auto& name : {"language", "prompt", "response_format", "stream",
-                             "temperature", "chunking_strategy"}) {
-        if (!single_field(name)) return;
-    }
-    const std::string format = req.form.has_field("response_format") ? req.form.get_field("response_format") : "json";
-    if (format != "json" && format != "text" && format != "verbose_json" &&
-        format != "srt" && format != "vtt") {
-        write_error(resp, 400, "unsupported_response_format",
-                    "response_format must be json, text, verbose_json, srt, or vtt",
-                    "response_format");
-        return;
-    }
-    if (req.form.has_field("stream")) {
-        const auto stream = req.form.get_field("stream");
-        if (stream != "true" && stream != "false") {
-            write_error(resp, 400, "invalid_transcription_request",
-                        "stream must be true or false", "stream");
-            return;
-        }
-        if (stream == "true") {
-            write_error(resp, 400, "unsupported_parameter",
-                        "streaming transcription is not supported by this native runtime",
-                        "stream");
-            return;
-        }
-    }
-    for (const auto& name : {"chunking_strategy", "include", "include[]",
-                             "keywords", "keywords[]", "known_speaker_names",
-                             "known_speaker_names[]", "known_speaker_references",
-                             "known_speaker_references[]", "languages",
-                             "languages[]"}) {
-        if (req.form.has_field(name)) {
-            write_error(resp, 400, "unsupported_parameter",
-                        std::string{name} +
-                            " is not supported by this native transcription runtime",
-                        name);
-            return;
-        }
-    }
-    std::vector<std::string> granularities =
-        req.form.get_fields("timestamp_granularities[]");
-    const auto plain_granularities =
-        req.form.get_fields("timestamp_granularities");
-    granularities.insert(granularities.end(), plain_granularities.begin(),
-                         plain_granularities.end());
-    if (!granularities.empty()) {
-        if (format != "verbose_json") {
-            write_error(resp, 400, "invalid_transcription_request",
-                        "timestamp_granularities requires verbose_json",
-                        "timestamp_granularities");
-            return;
-        }
-        if (granularities.size() != 1 || granularities.front() != "segment") {
-            write_error(resp, 400, "unsupported_parameter",
-                        "only segment timestamp granularity is supported",
-                        "timestamp_granularities");
-            return;
-        }
-    }
-    auto parameters = apply_transcription_parameters(
-        model::TranscriptionRequest{}, req.form);
-    if (!parameters) {
-        write_error(resp, 400, "invalid_transcription_request",
-                    parameters.error().message);
-        return;
-    }
-    const auto resolved_model = resolve_model_name(deps, model_name);
-    if (!resolved_model) {
-        write_error(resp, 404, "model_not_found", resolved_model.error().message);
-        return;
-    }
-    const std::string& runtime_model = resolved_model->resolved;
-    VoiceSessionGuard voice_session(req, deps);
-    auto job = begin_job(model_name, "audio_transcription");
-    if (deps.compatibility_profile == CompatibilityProfile::OpenAIDerivative) {
-        resp.set_header("X-InferDeck-Job-Id", std::to_string(job->id));
-    }
-    auto decode_permit = acquire_decode_permit(req, job);
-    if (!decode_permit) {
-        const int status = status_for(decode_permit.error().code);
-        const int internal_status = internal_status_for(decode_permit.error().code);
-        write_error(resp, status, "transcription_admission_failed",
-                    decode_permit.error().message);
-        record_media(deps, model_name, 0, internal_status, -1,
-                     0.0, 0, observation);
-        finish_job(job, internal_status == 499 ? "cancelled" : "failed");
-        return;
-    }
-    auto slot = acquire_media_slot(req, deps, runtime_model, job);
-    if (!slot) {
-        const int status = status_for(slot.error().code);
-        const int internal_status = internal_status_for(slot.error().code);
-        write_error(resp, status, "transcription_admission_failed",
-                    slot.error().message);
-        record_media(deps, model_name, 0, internal_status, -1,
-                     0.0, 0, observation);
-        finish_job(job, internal_status == 499 ? "cancelled" : "failed");
-        return;
-    }
-    SlotGuard guard{&deps.coordinator, runtime_model, *slot};
-    auto decoded = decode_audio(file.content);
-    decode_permit->release();
-    if (!decoded) { write_error(resp, 400, "invalid_audio", decoded.error().message); record_media(deps, model_name, 0, 400, *slot, 0.0, 0, observation); finish_job(job, "failed"); return; }
-    decoded->language = std::move(parameters->language);
-    decoded->prompt = std::move(parameters->prompt);
-    decoded->temperature = parameters->temperature;
-    const double input_audio_seconds =
-        static_cast<double>(decoded->pcm.size()) /
-        static_cast<double>(decoded->sample_rate);
-    auto result = deps.coordinator.transcribe(runtime_model, *slot, *decoded,
-        [&req, &deps, &model_name, job](int progress) {
-            update_job(job, progress);
-            if (deps.events) deps.events->publish("progress", nlohmann::json{{"id", job->id}, {"model", model_name}, {"modality", "audio_transcription"}, {"progress", progress}}.dump());
-            return !req.is_connection_closed() && !job->cancelled->load();
-        });
-    if (!result) {
-        const bool cancelled = job->cancelled->load() ||
-            result.error().code == foundation::ErrorCode::Cancelled;
-        const int status = cancelled ? 408 : status_for(result.error().code);
-        write_error(resp, status, "transcription_failed", result.error().message);
-        record_media(deps, model_name, 0, cancelled ? 499 : status, *slot,
-                     input_audio_seconds, 0, observation);
-        finish_job(job, cancelled ? "cancelled" : "failed");
-        return;
-    }
-    if (format == "text") {
-        resp.status = 200;
-        resp.set_content(result->text, "text/plain; charset=utf-8");
-    } else if (format == "verbose_json") {
-        write_json(resp, 200, verbose_transcription(*result, decoded->temperature));
-    } else if (format == "srt") {
-        resp.status = 200;
-        resp.set_content(subtitles(*result, false), "application/x-subrip; charset=utf-8");
-    } else if (format == "vtt") {
-        resp.status = 200;
-        resp.set_content(subtitles(*result, true), "text/vtt; charset=utf-8");
-    } else {
-        write_json(resp, 200, {{"text", result->text}});
-    }
-    record_media(deps, model_name, result->inference_ms, 200, *slot,
-                 input_audio_seconds, 0, observation);
-    finish_job(job, "completed");
-    voice_session.refresh_and_keep();
-}
+#include "transcription_routes.ipp"
 
 }
