+++ libs/llama_cpp_wrapper/src/llama_cpp_model.cpp
@@ -592,1294 +592,11 @@ bool fallback_tool_call_complete(const std::string& generated) {
 
 } // namespace
 
-std::string LlamaCppModel::version() {
-  const char* info = llama_print_system_info();
-  return info ? std::string(info) : std::string("unknown");
-}
+#include "llama_cpp_runtime.ipp"
 
-void LlamaCppModel::init_backend() {
-  if (!g_backend_initialized) {
-    llama_backend_init();
-    g_backend_initialized = true;
-    LOG_INFO("llama_backend_init", "Vulkan backend initialized");
-  }
-}
-void LlamaCppModel::shutdown_backend() {
-  if (g_backend_initialized) {
-    llama_backend_free();
-    g_backend_initialized = false;
-  }
-}
+#include "llama_cpp_embedding.ipp"
 
-LlamaCppModel::LlamaCppModel(inferdeck::model::ModelInfo info, LlamaCppConfig cfg)
-    : info_(std::move(info)), cfg_(std::move(cfg)) {
-  resolved_gguf_path_ = normalize_path(info_.gguf_path);
-  resolved_mmproj_path_ = normalize_path(info_.mmproj_path);
-}
-
-LlamaCppModel::~LlamaCppModel() {
-  // Stop scheduler before taking the mutex so no decode races with cleanup.
-  if (scheduler_) {
-    scheduler_->stop();
-    scheduler_.reset();
-  }
-  std::lock_guard lk(mtx_);
-  for (auto& s : slots_) s.busy = false;
-  slots_.clear();
-  if (speculative_) {
-    common_speculative_free(speculative_);
-    speculative_ = nullptr;
-  }
-  if (draft_ctx_) {
-    llama_free(draft_ctx_);
-    draft_ctx_ = nullptr;
-  }
-  if (shared_ctx_) {
-    llama_free(shared_ctx_);
-    shared_ctx_ = nullptr;
-  }
-  if (mtmd_) {
-    mtmd_free(mtmd_);
-    mtmd_ = nullptr;
-  }
-  if (chat_templates_) {
-    common_chat_templates_free(chat_templates_);
-    chat_templates_ = nullptr;
-  }
-  if (model_) {
-    llama_model_free(model_);
-    model_ = nullptr;
-  }
-  vocab_ = nullptr;
-  loaded_.store(false);
-}
-
-Result<void> LlamaCppModel::load() {
-  return load({});
-}
-
-Result<void> LlamaCppModel::load(
-    const inferdeck::model::LifecycleControl& control) {
-  std::lock_guard lk(mtx_);
-  if (loaded_.load()) return Result<void>{};
-  if (control.is_cancelled()) {
-    return Result<void>(std::unexpect,
-        make_error(ErrorCode::Cancelled, "model load cancelled"));
-  }
-  if (control.is_expired()) {
-    return Result<void>(std::unexpect,
-        make_error(ErrorCode::Timeout, "model load deadline expired"));
-  }
-  if (resolved_gguf_path_.empty()) {
-    return Result<void>(std::unexpect,
-        make_error(ErrorCode::NotFound, "empty gguf_path"));
-  }
-  std::error_code ec;
-  if (!std::filesystem::exists(resolved_gguf_path_, ec)) {
-    return Result<void>(std::unexpect,
-        make_error(ErrorCode::NotFound, "gguf not found: " + resolved_gguf_path_.string()));
-  }
-  if (info_.has_vision && resolved_mmproj_path_.empty()) {
-    return Result<void>(std::unexpect,
-        make_error(ErrorCode::NotFound, "vision model has no mmproj_path"));
-  }
-  if (info_.has_vision && !std::filesystem::exists(resolved_mmproj_path_, ec)) {
-    return Result<void>(std::unexpect,
-        make_error(ErrorCode::NotFound, "mmproj not found: " + resolved_mmproj_path_.string()));
-  }
-
-  llama_model_params mparams = llama_model_default_params();
-  mparams.load_mode = cfg_.use_mmap
-      ? (cfg_.use_mlock ? LLAMA_LOAD_MODE_MMAP_MLOCK : LLAMA_LOAD_MODE_MMAP)
-      : (cfg_.use_mlock ? LLAMA_LOAD_MODE_MLOCK : LLAMA_LOAD_MODE_NONE);
-  mparams.n_gpu_layers = cfg_.n_gpu_layers.value_or(-1);
-  mparams.load_mtp = cfg_.mtp_enabled;
-  mparams.progress_callback = [](float, void* user_data) {
-    const auto* lifecycle = static_cast<
-        const inferdeck::model::LifecycleControl*>(user_data);
-    return !lifecycle->is_cancelled() && !lifecycle->is_expired();
-  };
-  mparams.progress_callback_user_data =
-      const_cast<inferdeck::model::LifecycleControl*>(&control);
-
-  llama_backend_init();
-  const char* sys_info = llama_print_system_info();
-  if (sys_info) {
-    LOG_INFO("llama_system_info", "{}", sys_info);
-  }
-  LOG_INFO("llama_model_load_config",
-           "model={} path={} use_mmap={} use_mlock={} n_gpu_layers={} n_ctx={} n_slots={} n_batch={} n_ubatch={} flash_attn={} kv_offload={} op_offload={} cache_type_k={} cache_type_v={} mtp_enabled={} mtp_draft_tokens={} mtp_max_active_requests={} swa_full={}",
-           info_.name,
-           resolved_gguf_path_.string(),
-           cfg_.use_mmap,
-           cfg_.use_mlock,
-           mparams.n_gpu_layers,
-           info_.context_size,
-           info_.n_slots,
-           cfg_.n_batch,
-           cfg_.n_ubatch,
-           cfg_.flash_attn,
-           cfg_.kv_offload,
-           cfg_.op_offload,
-           cfg_.cache_type_k,
-           cfg_.cache_type_v,
-           cfg_.mtp_enabled,
-           cfg_.mtp_draft_tokens,
-           cfg_.mtp_max_active_requests,
-           cfg_.swa_full);
-  log_memory_snapshot("llama_model_load_memory_before", info_.name);
-
-  model_ = llama_model_load_from_file(resolved_gguf_path_.string().c_str(), mparams);
-  if (model_ == nullptr) {
-    if (control.is_cancelled()) {
-      return Result<void>(std::unexpect,
-          make_error(ErrorCode::Cancelled, "model load cancelled"));
-    }
-    if (control.is_expired()) {
-      return Result<void>(std::unexpect,
-          make_error(ErrorCode::Timeout, "model load deadline expired"));
-    }
-    const char* err = llama_print_system_info();
-    LOG_ERROR("model_load_failed", "llama_model_load_from_file returned null for {}", resolved_gguf_path_.string());
-    if (err) LOG_ERROR("model_load_failed", "system_info: {}", err);
-    return Result<void>(std::unexpect,
-        make_error(ErrorCode::Internal,
-                   "llama_model_load_from_file returned null for " + resolved_gguf_path_.string()));
-  }
-  log_memory_snapshot("llama_model_loaded_memory_after", info_.name);
-  vocab_ = llama_model_get_vocab(model_);
-  if (vocab_ == nullptr) {
-    llama_model_free(model_);
-    model_ = nullptr;
-    return Result<void>(std::unexpect,
-        make_error(ErrorCode::ParseError, "llama_model_get_vocab returned null"));
-  }
-  // Empty cfg_.chat_template => use the template embedded in the GGUF; a non-empty
-  // value is a literal Jinja override (e.g. the corrected Qwen3.6 template that avoids
-  // the "No user query found in messages." crash during multi-step tool calling).
-  if (info_.supports("chat_completions")) {
-    chat_templates_ = common_chat_templates_init(model_, cfg_.chat_template).release();
-    if (chat_templates_ == nullptr) {
-      llama_model_free(model_);
-      model_ = nullptr;
-      return Result<void>(std::unexpect,
-          make_error(ErrorCode::ParseError, "common_chat_templates_init returned null"));
-    }
-  }
-  if (info_.has_vision) {
-    auto mtmd_params = mtmd_context_params_default();
-    mtmd_params.use_gpu = true;
-    mtmd_params.n_threads = cfg_.n_threads;
-    mtmd_params.flash_attn_type = flash_attn_from_string(cfg_.flash_attn);
-    mtmd_ = mtmd_init_from_file(
-        resolved_mmproj_path_.string().c_str(), model_, mtmd_params);
-    if (mtmd_ == nullptr || !mtmd_support_vision(mtmd_)) {
-      if (mtmd_) {
-        mtmd_free(mtmd_);
-        mtmd_ = nullptr;
-      }
-      if (chat_templates_) {
-        common_chat_templates_free(chat_templates_);
-        chat_templates_ = nullptr;
-      }
-      llama_model_free(model_);
-      model_ = nullptr;
-      vocab_ = nullptr;
-      return Result<void>(std::unexpect,
-          make_error(ErrorCode::ParseError,
-                     "failed to load vision projector: " + resolved_mmproj_path_.string()));
-    }
-    LOG_INFO("vision_projector_loaded",
-             "model={} path={}", info_.name, resolved_mmproj_path_.string());
-  }
-  auto ctx_res = init_shared_context_locked();
-  if (!ctx_res.has_value()) {
-    if (speculative_) {
-      common_speculative_free(speculative_);
-      speculative_ = nullptr;
-    }
-    if (draft_ctx_) { llama_free(draft_ctx_); draft_ctx_ = nullptr; }
-    if (shared_ctx_) { llama_free(shared_ctx_); shared_ctx_ = nullptr; }
-    if (mtmd_) { mtmd_free(mtmd_); mtmd_ = nullptr; }
-    slots_.clear();
-    if (chat_templates_) {
-      common_chat_templates_free(chat_templates_);
-      chat_templates_ = nullptr;
-    }
-    llama_model_free(model_);
-    model_ = nullptr;
-    vocab_ = nullptr;
-    return ctx_res;
-  }
-  // Populate chat_template_meta_ once from the model's Jinja template.
-  if (info_.supports("chat_completions")) {
-    InferenceRequest dummy;
-    dummy.messages.push_back({"user", "hello"});
-    auto meta_res = apply_chat_template(dummy);
-    if (meta_res.has_value()) chat_template_meta_ = std::move(meta_res->meta);
-  }
-  loaded_.store(true);
-  log_memory_snapshot("llama_contexts_initialized_memory_after", info_.name);
-  LOG_INFO("chat_template_loaded", "model={} kind=jinja", info_.name);
-  return Result<void>{};
-}
-
-Result<void> LlamaCppModel::init_shared_context_locked() {
-  // One shared context for all slots.
-  // n_ctx = context_size * n_slots so each slot gets its own context window via sequence IDs.
-  // n_seq_max = n_slots so the KV cache can track each slot's sequence independently.
-  const int n_slots = std::max(1, info_.n_slots);
-  const int ctx_per_slot = std::max(512, info_.context_size);
-  const int total_ctx = ctx_per_slot * n_slots;
-
-  llama_context_params cparams = llama_context_default_params();
-  cparams.n_ctx      = static_cast<std::uint32_t>(total_ctx);
-  cparams.n_seq_max  = static_cast<std::uint32_t>(n_slots);
-  cparams.n_threads  = cfg_.n_threads;
-  cparams.n_batch    = static_cast<std::uint32_t>(std::max(1, cfg_.n_batch));
-  cparams.n_ubatch   = static_cast<std::uint32_t>(std::max(1, cfg_.n_ubatch));
-  cparams.flash_attn_type = flash_attn_from_string(cfg_.flash_attn);
-  cparams.offload_kqv = cfg_.kv_offload;
-  cparams.op_offload  = cfg_.op_offload;
-  cparams.swa_full    = cfg_.swa_full;
-  cparams.embeddings  = info_.supports("embeddings");
-  cparams.type_k      = cache_type_from_string(cfg_.cache_type_k);
-  cparams.type_v      = cache_type_from_string(cfg_.cache_type_v);
-  cparams.n_rs_seq    = cfg_.mtp_enabled
-      ? static_cast<std::uint32_t>(std::max(1, cfg_.mtp_draft_tokens))
-      : 0;
-
-  LOG_INFO("llama_shared_context_config",
-           "model={} n_slots={} ctx_per_slot={} total_ctx={} n_seq_max={} "
-           "n_batch={} n_ubatch={} flash_attn={} kv_offload={} op_offload={} "
-           "cache_type_k={} cache_type_v={} mtp_enabled={} mtp_draft_tokens={} mtp_max_active_requests={} swa_full={}",
-           info_.name, n_slots, ctx_per_slot, total_ctx, n_slots,
-           cparams.n_batch, cparams.n_ubatch,
-           cfg_.flash_attn, cfg_.kv_offload, cfg_.op_offload,
-           cfg_.cache_type_k, cfg_.cache_type_v,
-           cfg_.mtp_enabled, cfg_.mtp_draft_tokens,
-           cfg_.mtp_max_active_requests, cfg_.swa_full);
-
-  shared_ctx_ = llama_init_from_model(model_, cparams);
-  if (shared_ctx_ == nullptr) {
-    return Result<void>(std::unexpect,
-        make_error(ErrorCode::OutOfMemory,
-                   "llama_init_from_model returned null (shared context, total_ctx=" +
-                   std::to_string(total_ctx) + ")"));
-  }
-
-  auto draft_seq_rm_type = COMMON_CONTEXT_SEQ_RM_TYPE_NO;
-  if (cfg_.mtp_enabled) {
-    auto draft_params = cparams;
-    draft_params.ctx_type = LLAMA_CONTEXT_TYPE_MTP;
-    draft_params.n_rs_seq = 0;
-    draft_params.n_ubatch = std::min<std::uint32_t>(
-        draft_params.n_ubatch, 512);
-    draft_ctx_ = llama_init_from_model(model_, draft_params);
-    if (draft_ctx_ == nullptr) {
-      return Result<void>(std::unexpect,
-          make_error(ErrorCode::InvalidArgument,
-                     "MTP is enabled but the GGUF has no usable MTP head"));
-    }
-
-    (void)common_context_can_seq_rm(shared_ctx_);
-    draft_seq_rm_type = common_context_can_seq_rm(draft_ctx_);
-    if (draft_seq_rm_type == COMMON_CONTEXT_SEQ_RM_TYPE_NO) {
-      return Result<void>(std::unexpect,
-          make_error(ErrorCode::InvalidArgument,
-                     "MTP draft context does not support sequence removal"));
-    }
-
-    common_params_speculative params;
-    params.types = {COMMON_SPECULATIVE_TYPE_DRAFT_MTP};
-    params.draft.n_max = std::clamp(cfg_.mtp_draft_tokens, 1, 4);
-    params.draft.n_min = 0;
-    params.draft.p_min = std::clamp(cfg_.mtp_p_min, 0.0f, 1.0f);
-    params.draft.cache_type_k =
-        cache_type_from_string(cfg_.cache_type_k);
-    params.draft.cache_type_v =
-        cache_type_from_string(cfg_.cache_type_v);
-    params.draft.ctx_tgt = shared_ctx_;
-    params.draft.ctx_dft = draft_ctx_;
-    try {
-      speculative_ = common_speculative_init(
-          params, static_cast<std::uint32_t>(n_slots));
-    } catch (const std::exception& error) {
-      return Result<void>(std::unexpect,
-          make_error(ErrorCode::Internal,
-                     std::string("MTP initialization failed: ") +
-                     error.what()));
-    }
-    if (speculative_ == nullptr) {
-      return Result<void>(std::unexpect,
-          make_error(ErrorCode::Internal,
-                     "MTP initialization returned null"));
-    }
-    LOG_INFO("llama_mtp_initialized",
-             "model={} draft_tokens={} p_min={} max_active_requests={} draft_seq_rm_type={}",
-             info_.name,
-             params.draft.n_max,
-             params.draft.p_min,
-             cfg_.mtp_max_active_requests,
-             static_cast<int>(draft_seq_rm_type));
-  }
-
-  slots_.clear();
-  slots_.resize(n_slots);
-
-  // Spawn the scheduler that owns the decode loop for this context.
-  if (info_.supports("chat_completions")) {
-    scheduler_ = std::make_unique<ContinuousBatchScheduler>(
-        shared_ctx_,
-        draft_ctx_,
-        speculative_,
-        mtmd_,
-        model_,
-        vocab_,
-        cfg_.n_batch,
-        cfg_.mtp_max_active_requests,
-        draft_seq_rm_type);
-  }
-
-  return Result<void>{};
-}
-
-Result<void> LlamaCppModel::unload() {
-  // Stop the scheduler first (joins its thread) so no decode can race with teardown.
-  if (scheduler_) {
-    scheduler_->stop();
-    scheduler_.reset();
-  }
-  std::lock_guard lk(mtx_);
-  if (!loaded_.load()) return Result<void>{};
-  log_memory_snapshot("llama_model_unload_memory_before", info_.name);
-  for (auto& s : slots_) s.busy = false;
-  slots_.clear();
-  if (speculative_) {
-    common_speculative_free(speculative_);
-    speculative_ = nullptr;
-  }
-  if (draft_ctx_) {
-    llama_free(draft_ctx_);
-    draft_ctx_ = nullptr;
-  }
-  if (shared_ctx_) {
-    llama_free(shared_ctx_);
-    shared_ctx_ = nullptr;
-  }
-  if (mtmd_) {
-    mtmd_free(mtmd_);
-    mtmd_ = nullptr;
-  }
-  if (chat_templates_) {
-    common_chat_templates_free(chat_templates_);
-    chat_templates_ = nullptr;
-  }
-  if (model_) {
-    llama_model_free(model_);
-    model_ = nullptr;
-  }
-  vocab_ = nullptr;
-  loaded_.store(false);
-  log_memory_snapshot("llama_model_unload_memory_after", info_.name);
-  return Result<void>{};
-}
-
-int LlamaCppModel::vram_usage_mb() const noexcept {
-  return estimate_vram_mb(info_.n_slots);
-}
-
-bool LlamaCppModel::can_resize_slots() const noexcept {
-  return info_.vram_fixed_mb > 0 && info_.vram_per_slot_mb > 0 &&
-         info_.n_slots > info_.min_slots;
-}
-
-int LlamaCppModel::estimate_vram_mb(int slots) const noexcept {
-  if (info_.vram_fixed_mb > 0 && info_.vram_per_slot_mb > 0) {
-    return info_.vram_fixed_mb + info_.vram_per_slot_mb * std::max(info_.min_slots, slots);
-  }
-  return info_.vram_required_mb;
-}
-
-Result<void> LlamaCppModel::resize_slots(int slots) {
-  if (slots < info_.min_slots || slots > info_.n_slots) {
-    return Result<void>(std::unexpect,
-        make_error(ErrorCode::InvalidArgument, "invalid slot capacity: " + std::to_string(slots)));
-  }
-  if (slots == info_.n_slots) return Result<void>{};
-  {
-    std::lock_guard lk(mtx_);
-    if (std::any_of(slots_.begin(), slots_.end(), [](const SlotState& slot) { return slot.busy; })) {
-      return Result<void>(std::unexpect,
-          make_error(ErrorCode::Unavailable, "cannot resize while slots are active"));
-    }
-  }
-  const int previous = info_.n_slots;
-  const bool was_loaded = loaded_.load();
-  if (was_loaded) {
-    auto unloaded = unload();
-    if (!unloaded) return unloaded;
-  }
-  info_.n_slots = slots;
-  if (!was_loaded) return Result<void>{};
-  auto loaded = load();
-  if (loaded) return loaded;
-  info_.n_slots = previous;
-  (void)load();
-  return loaded;
-}
-
-int LlamaCppModel::n_free_slots() const noexcept {
-  std::lock_guard lk(mtx_);
-  int free = 0;
-  for (const auto& s : slots_) if (!s.busy) ++free;
-  return free;
-}
-
-Result<int> LlamaCppModel::acquire_slot() {
-  std::lock_guard lk(mtx_);
-  if (!loaded_.load()) {
-    return Result<int>(std::unexpect,
-        make_error(ErrorCode::Internal, "model not loaded"));
-  }
-  for (int i = 0; i < static_cast<int>(slots_.size()); ++i) {
-    if (!slots_[i].busy) {
-      slots_[i].busy = true;
-      return Result<int>(i);
-    }
-  }
-  return Result<int>(std::unexpect, make_error(ErrorCode::Unavailable, "no free slots"));
-}
-
-Result<void> LlamaCppModel::release_slot(int slot_id) {
-  std::lock_guard lk(mtx_);
-  if (slot_id < 0 || slot_id >= static_cast<int>(slots_.size())) {
-    return Result<void>(std::unexpect,
-        make_error(ErrorCode::InvalidArgument,
-                   "slot_id out of range: " + std::to_string(slot_id)));
-  }
-  auto& slot = slots_[slot_id];
-  slot.busy = false;
-  return Result<void>{};
-}
-
-bool LlamaCppModel::slot_busy(int slot_id) const noexcept {
-  std::lock_guard lk(mtx_);
-  if (slot_id < 0 || slot_id >= static_cast<int>(slots_.size())) return false;
-  return slots_[slot_id].busy;
-}
-
-Result<void> LlamaCppModel::reset_all_slots() noexcept {
-  std::lock_guard lk(mtx_);
-  if (shared_ctx_) {
-    // Clear KV entries for every slot's sequence individually.
-    // This avoids llama_memory_clear which would also clear non-slot sequences.
-    auto* mem = llama_get_memory(shared_ctx_);
-    if (mem) {
-      for (int i = 0; i < static_cast<int>(slots_.size()); ++i) {
-        llama_memory_seq_rm(mem, i, 0, -1);
-      }
-    }
-  }
-  if (draft_ctx_) {
-    auto* mem = llama_get_memory(draft_ctx_);
-    if (mem) {
-      for (int i = 0; i < static_cast<int>(slots_.size()); ++i) {
-        llama_memory_seq_rm(mem, i, 0, -1);
-      }
-    }
-  }
-  for (auto& s : slots_) {
-    s.busy = false;
-    s.last_prompt_tokens.clear();
-    s.recurrent_checkpoint.reset();
-    s.recurrent_draft_checkpoint.reset();
-    s.recurrent_mtp_checkpoint.reset();
-    s.checkpoint_pos = 0;
-    s.mtp_cache_synced = true;
-  }
-  return Result<void>{};
-}
-
-Result<inferdeck::model::EmbeddingResult> LlamaCppModel::embed(
-    int slot_id, const inferdeck::model::EmbeddingRequest& request,
-    const std::function<bool()>& cancelled) {
-  const auto started = std::chrono::steady_clock::now();
-  std::lock_guard lk(mtx_);
-  if (!loaded_.load() || !shared_ctx_ || !model_ || !vocab_) {
-    return Result<EmbeddingResult>(std::unexpect,
-        make_error(ErrorCode::NotLoaded, "embedding model not loaded"));
-  }
-  if (!info_.supports("embeddings") || scheduler_) {
-    return Result<EmbeddingResult>(std::unexpect,
-        make_error(ErrorCode::InvalidArgument, "dedicated embedding model required"));
-  }
-  if (slot_id < 0 || slot_id >= static_cast<int>(slots_.size()) || !slots_[slot_id].busy) {
-    return Result<EmbeddingResult>(std::unexpect,
-        make_error(ErrorCode::InvalidArgument, "embedding slot is not acquired"));
-  }
-
-  EmbeddingResult result;
-  const int native_dimensions = llama_model_n_embd_out(model_);
-  if (request.dimensions && *request.dimensions != native_dimensions) {
-    return Result<EmbeddingResult>(std::unexpect,
-        make_error(ErrorCode::InvalidArgument,
-                   "requested dimensions must equal native dimensions: " +
-                   std::to_string(native_dimensions)));
-  }
-  result.embeddings.reserve(request.inputs.size());
-  for (const auto& input : request.inputs) {
-    if (cancelled && cancelled()) {
-      return Result<EmbeddingResult>(std::unexpect,
-          make_error(ErrorCode::Cancelled, "embedding request cancelled"));
-    }
-    std::vector<llama_token> tokens;
-    if (const auto* text = std::get_if<inferdeck::model::EmbeddingTextInput>(&input)) {
-      tokens = common_tokenize(vocab_, text->text, true, true);
-    } else {
-      const auto& token_input = std::get<inferdeck::model::EmbeddingTokenInput>(input);
-      const int32_t vocabulary_size = llama_vocab_n_tokens(vocab_);
-      tokens.reserve(token_input.tokens.size());
-      for (const std::int32_t token : token_input.tokens) {
-        if (token < 0 || token >= vocabulary_size) {
-          return Result<EmbeddingResult>(std::unexpect,
-              make_error(ErrorCode::InvalidArgument,
-                         "embedding input contains an invalid token ID"));
-        }
-        tokens.push_back(static_cast<llama_token>(token));
-      }
-    }
-    if (tokens.empty() || tokens.size() > llama_n_batch(shared_ctx_)) {
-      return Result<EmbeddingResult>(std::unexpect,
-          make_error(ErrorCode::InvalidArgument, "embedding input exceeds model batch limit"));
-    }
-    llama_batch batch = llama_batch_init(static_cast<int32_t>(tokens.size()), 0, 1);
-    for (std::size_t i = 0; i < tokens.size(); ++i) {
-      common_batch_add(batch, tokens[i], static_cast<llama_pos>(i), {0}, true);
-    }
-    llama_memory_clear(llama_get_memory(shared_ctx_), true);
-    if (llama_decode(shared_ctx_, batch) < 0) {
-      llama_batch_free(batch);
-      return Result<EmbeddingResult>(std::unexpect,
-          make_error(ErrorCode::Internal, "embedding decode failed"));
-    }
-    const float* raw = llama_pooling_type(shared_ctx_) == LLAMA_POOLING_TYPE_NONE
-        ? llama_get_embeddings_ith(shared_ctx_, batch.n_tokens - 1)
-        : llama_get_embeddings_seq(shared_ctx_, 0);
-    if (!raw) {
-      llama_batch_free(batch);
-      return Result<EmbeddingResult>(std::unexpect,
-          make_error(ErrorCode::Internal, "embedding output unavailable"));
-    }
-    auto& output = result.embeddings.emplace_back(static_cast<std::size_t>(native_dimensions));
-    common_embd_normalize(raw, output.data(), native_dimensions, 2);
-    result.prompt_tokens += static_cast<int>(tokens.size());
-    llama_batch_free(batch);
-  }
-  result.duration_ms = std::chrono::duration<float, std::milli>(
-      std::chrono::steady_clock::now() - started).count();
-  return Result<EmbeddingResult>(std::move(result));
-}
-
-Result<ChatTemplateResult> LlamaCppModel::apply_chat_template(
-    const InferenceRequest& req, int max_prompt_tokens) {
-  if (!chat_templates_) {
-    return Result<ChatTemplateResult>(std::unexpect, make_error(ErrorCode::Internal, "chat templates not initialized"));
-  }
-
-  auto caps = common_chat_templates_get_caps(chat_templates_);
-  LlamaChatAdapterOptions adapter_options;
-  adapter_options.supports_thinking =
-      common_chat_templates_support_enable_thinking(chat_templates_);
-  adapter_options.supports_parallel_tool_calls =
-      caps["supports_parallel_tool_calls"];
-  adapter_options.default_reasoning_format = info_.reasoning_format.empty()
-      ? cfg_.reasoning_format : info_.reasoning_format;
-  auto adapted = adapt_generation_request(req, info_, adapter_options);
-  if (!adapted) {
-    return Result<ChatTemplateResult>(std::unexpect,
-        make_error(adapted.error().code, adapted.error().message));
-  }
-  auto inputs = std::move(adapted->inputs);
-  auto media = std::move(adapted->media);
-  if (!media.empty() && mtmd_ == nullptr) {
-    return Result<ChatTemplateResult>(std::unexpect,
-        make_error(ErrorCode::InvalidArgument,
-                   "image input requires a loaded vision projector"));
-  }
-
-  {
-    std::size_t sys_chars = 0, user_chars = 0, tool_result_chars = 0;
-    int n_sys = 0, n_user = 0, n_assistant = 0, n_tool = 0;
-    for (const auto& m : inputs.messages) {
-      const std::size_t len = m.content.size();
-      if (m.role == "system")    { ++n_sys; sys_chars += len; }
-      else if (m.role == "user") { ++n_user; user_chars += len; }
-      else if (m.role == "assistant") { ++n_assistant; }
-      else if (m.role == "tool") { ++n_tool; tool_result_chars += len; }
-    }
-    LOG_INFO("request_shape",
-             "model={} msgs={} [sys={} sys_chars={} user={} asst={} tool_results={} tool_result_chars={}] "
-             "tools_defined={} max_tokens={}",
-             info_.name, inputs.messages.size(),
-             n_sys, sys_chars, n_user, n_assistant, n_tool, tool_result_chars,
-             req.tools.size(), req.max_output_tokens);
-  }
-
-  if (req.reasoning_effort) {
-    LOG_INFO("reasoning_effort_applied", "model={} effort={}",
-             info_.name, *req.reasoning_effort);
-  }
-
-  try {
-  auto chat_params = common_chat_templates_apply(chat_templates_, inputs);
-
-  // History-aware truncation (issue #38): rather than middle-dropping the raw
-  // token stream (which severs conversation history and defeats KV prefix
-  // reuse), drop the oldest *whole* non-system messages and re-template until
-  // the prompt fits the budget. The leading system block and the most recent
-  // turn are always preserved.
-  if (max_prompt_tokens > 0 && media.empty()) {
-    auto count_prompt_tokens = [&](const std::string& p) -> int {
-      if (p.empty()) return 0;
-      const int n = llama_tokenize(vocab_, p.data(), static_cast<int>(p.size()),
-                                   nullptr, 0, llama_vocab_get_add_bos(vocab_), true);
-      return n < 0 ? -n : n;
-    };
-    std::size_t sys_end = 0;
-    while (sys_end < inputs.messages.size() && inputs.messages[sys_end].role == "system")
-      ++sys_end;
-    int dropped = 0;
-    while (count_prompt_tokens(chat_params.prompt) >= max_prompt_tokens &&
-           inputs.messages.size() - sys_end > 1) {
-      inputs.messages.erase(inputs.messages.begin() + static_cast<std::ptrdiff_t>(sys_end));
-      ++dropped;
-      // Drop any now-orphaned tool results whose assistant tool_call was removed.
-      while (inputs.messages.size() - sys_end > 1 &&
-             inputs.messages[sys_end].role == "tool") {
-        inputs.messages.erase(inputs.messages.begin() + static_cast<std::ptrdiff_t>(sys_end));
-        ++dropped;
-      }
-      chat_params = common_chat_templates_apply(chat_templates_, inputs);
-    }
-    if (dropped > 0) {
-      LOG_WARN("chat_history_truncated",
-               "model={} dropped_messages={} kept_messages={} prompt_tokens={} budget={}",
-               info_.name, dropped, inputs.messages.size(),
-               count_prompt_tokens(chat_params.prompt), max_prompt_tokens);
-    }
-  }
-
-  ChatTemplateMeta meta;
-  meta.thinking_start_tag = chat_params.thinking_start_tag;
-  meta.thinking_end_tags = chat_params.thinking_end_tags;
-  meta.preserved_tokens = chat_params.preserved_tokens;
-  meta.supports_thinking = chat_params.supports_thinking;
-
-  ChatTemplateResult result;
-  result.prompt = chat_params.prompt;
-  result.media = std::move(media);
-  result.stop_strings = chat_params.additional_stops;
-  result.parser_params = common_chat_parser_params(chat_params);
-  result.parser_params.reasoning_format = inputs.reasoning_format;
-  result.parser_params.reasoning_in_content = false;
-  result.parser_params.parse_tool_calls = !inputs.tools.empty() && inputs.tool_choice != COMMON_CHAT_TOOL_CHOICE_NONE;
-  if (!chat_params.parser.empty()) {
-    result.parser_params.parser.load(chat_params.parser);
-  }
-  result.stop_strings.insert(result.stop_strings.end(), req.stop.begin(),
-                             req.stop.end());
-  // Sampler params: explicit per-request values win; otherwise
-  // fall back to the server-side SamplingConfig defaults (issue #42), which
-  // mirror stock llama-server (DRY off, repeat_penalty neutral).
-  const auto& sc = cfg_.sampling;
-  result.sampling_params.temp          = req.sampling.temperature.value_or(sc.temperature);
-  result.sampling_params.top_p         = req.sampling.top_p.value_or(sc.top_p);
-  result.sampling_params.top_k         = req.sampling.top_k.value_or(sc.top_k);
-  result.sampling_params.min_p         = req.sampling.min_p.value_or(sc.min_p);
-  result.sampling_params.penalty_repeat = req.sampling.repeat_penalty.value_or(sc.repeat_penalty);
-  result.sampling_params.penalty_last_n = req.sampling.repeat_last_n.value_or(sc.repeat_last_n);
-  result.sampling_params.dry_multiplier     = sc.dry_multiplier;
-  result.sampling_params.dry_base           = sc.dry_base;
-  result.sampling_params.dry_allowed_length = sc.dry_allowed_length;
-  result.sampling_params.dry_penalty_last_n = sc.dry_penalty_last_n;
-  result.sampling_params.dry_sequence_breakers = sc.dry_seq_breakers;
-  result.sampling_params.seed = req.sampling.seed >= 0
-      ? static_cast<std::uint32_t>(req.sampling.seed) : LLAMA_DEFAULT_SEED;
-
-  // DEBUG (issue #42 diagnosis): log what the client sent vs what was resolved,
-  // so we can see whether OpenCode/Claude Code override the server-side config.
-  auto opt_f = [](const std::optional<float>& v) {
-    return v.has_value() ? std::to_string(*v) : std::string("unset");
-  };
-  auto opt_i = [](const std::optional<int>& v) {
-    return v.has_value() ? std::to_string(*v) : std::string("unset");
-  };
-  LOG_INFO("sampling_resolved",
-           "model={} client[temp={} top_p={} top_k={} repeat_penalty={} repeat_last_n={}] "
-           "resolved[temp={:.3f} top_p={:.3f} top_k={} min_p={:.3f} repeat_penalty={:.3f} "
-           "repeat_last_n={} dry_mult={:.3f}]",
-           info_.name, opt_f(req.sampling.temperature), opt_f(req.sampling.top_p),
-           opt_i(req.sampling.top_k), opt_f(req.sampling.repeat_penalty),
-           opt_i(req.sampling.repeat_last_n),
-           result.sampling_params.temp, result.sampling_params.top_p,
-           result.sampling_params.top_k, result.sampling_params.min_p,
-           result.sampling_params.penalty_repeat, result.sampling_params.penalty_last_n,
-           result.sampling_params.dry_multiplier);
-
-  if (!chat_params.grammar.empty()) {
-    if (!inputs.tools.empty() && inputs.tool_choice != COMMON_CHAT_TOOL_CHOICE_NONE) {
-      result.sampling_params.grammar = {COMMON_GRAMMAR_TYPE_TOOL_CALLS, chat_params.grammar};
-    } else if (!inputs.json_schema.empty()) {
-      result.sampling_params.grammar = {COMMON_GRAMMAR_TYPE_OUTPUT_FORMAT, chat_params.grammar};
-    } else {
-      result.sampling_params.grammar = {COMMON_GRAMMAR_TYPE_USER, chat_params.grammar};
-    }
-  }
-  result.sampling_params.grammar_lazy = chat_params.grammar_lazy;
-  result.sampling_params.grammar_triggers = chat_params.grammar_triggers;
-  result.sampling_params.generation_prompt = chat_params.generation_prompt;
-  for (const auto& token_str : chat_params.preserved_tokens) {
-    auto toks = tokenize_stop_strings(vocab_, {token_str});
-    for (auto t : toks) result.sampling_params.preserved_tokens.insert(t);
-  }
-  result.meta = std::move(meta);
-
-  return Result<ChatTemplateResult>(std::move(result));
-  } catch (const std::exception& e) {
-    LOG_WARN("chat_template_failed", "model={} error={}", info_.name, e.what());
-    return Result<ChatTemplateResult>(std::unexpect,
-        make_error(ErrorCode::ParseError, std::string("chat template failed: ") + e.what()));
-  }
-}
-
-// Tokenizes the request, checks context limits, snapshots per-slot KV state,
-// and initialises a sampler. All of this runs on the HTTP handler thread
-// before the task is handed off to the scheduler.
-Result<LlamaCppModel::PredictSetup> LlamaCppModel::prepare_inference(
-    int slot_id, const InferenceRequest& req) {
-  PredictSetup s;
-  // Compute the prompt-token budget so apply_chat_template can drop whole
-  // oldest messages (history-aware truncation, issue #38) before tokenizing.
-  // Mirrors the reserve/target maths in maybe_truncate_prompt, which remains as
-  // a hard safety net for the pathological single-oversized-message case.
-  const int n_ctx_seq = static_cast<int>(llama_n_ctx_seq(shared_ctx_));
-  int budget = 0;
-  if (cfg_.truncate_prompt && n_ctx_seq > 0) {
-    // See maybe_truncate_prompt: clamp bounds must satisfy lo <= hi (UB
-    // otherwise) when n_ctx_seq < 1024.
-    const int reserve_hi = n_ctx_seq / 4;
-    const int reserve = std::clamp(req.max_output_tokens > 0 ? req.max_output_tokens : 1024,
-                                   std::min(256, reserve_hi), reserve_hi);
-    budget = n_ctx_seq - reserve - 1;
-  }
-  auto tmpl_res = apply_chat_template(req, budget);
-  if (!tmpl_res.has_value())
-    return Result<PredictSetup>(std::unexpect, tmpl_res.error());
-
-  s.parser_params   = std::move(tmpl_res->parser_params);
-  s.sampling_params = std::move(tmpl_res->sampling_params);
-  s.stop_strings    = std::move(tmpl_res->stop_strings);
-  s.stop_tokens     = tokenize_stop_strings(vocab_, s.stop_strings);
-
-  const std::string& prompt = tmpl_res->prompt;
-  const bool add_bos = llama_vocab_get_add_bos(vocab_);
-  const bool has_media = !tmpl_res->media.empty();
-  int n_tokens = 0;
-  if (has_media) {
-    if (mtmd_ == nullptr) {
-      return Result<PredictSetup>(std::unexpect,
-          make_error(ErrorCode::InvalidArgument,
-                     "image input requires a loaded vision projector"));
-    }
-    mtmd::bitmaps bitmaps;
-    for (const auto& data : tmpl_res->media) {
-      auto decoded = mtmd_helper_bitmap_init_from_buf(
-          mtmd_, data.data(), data.size(), false);
-      mtmd::bitmap bitmap(decoded.bitmap);
-      mtmd_helper::video_ptr video(decoded.video_ctx);
-      if (!bitmap.ptr) {
-        return Result<PredictSetup>(std::unexpect,
-            make_error(ErrorCode::InvalidArgument,
-                       "image payload could not be decoded"));
-      }
-      if (video) {
-        return Result<PredictSetup>(std::unexpect,
-            make_error(ErrorCode::InvalidArgument,
-                       "only image media is supported for chat inference"));
-      }
-      bitmaps.entries.push_back(std::move(bitmap));
-    }
-    mtmd_input_text input{
-        prompt.c_str(),
-        prompt.size(),
-        add_bos,
-        true,
-    };
-    mtmd::input_chunks chunks(mtmd_input_chunks_init());
-    auto bitmap_ptrs = bitmaps.c_ptr();
-    const auto vision_tokenize_started = std::chrono::steady_clock::now();
-    const int tokenized = mtmd_tokenize(
-        mtmd_, chunks.ptr.get(), &input,
-        bitmap_ptrs.data(), bitmap_ptrs.size());
-    if (tokenized != 0) {
-      return Result<PredictSetup>(std::unexpect,
-          make_error(ErrorCode::InvalidArgument,
-                     "multimodal prompt tokenization failed (rc=" +
-                     std::to_string(tokenized) + ")"));
-    }
-    LOG_INFO("vision_prompt_tokenized",
-             "model={} images={} chunks={} tokens={} positions={} duration_ms={:.3f}",
-             info_.name,
-             tmpl_res->media.size(),
-             chunks.size(),
-             mtmd_helper_get_n_tokens(chunks.ptr.get()),
-             mtmd_helper_get_n_pos(chunks.ptr.get()),
-             std::chrono::duration<float, std::milli>(
-                 std::chrono::steady_clock::now() - vision_tokenize_started).count());
-    s.prompt_position_count = static_cast<int>(
-        mtmd_helper_get_n_pos(chunks.ptr.get()));
-    for (std::size_t index = 0; index < chunks.size(); ++index) {
-      const auto* chunk = chunks[index];
-      const auto type = mtmd_input_chunk_get_type(chunk);
-      if (type == MTMD_INPUT_CHUNK_TYPE_TEXT) {
-        std::size_t count = 0;
-        const auto* tokens = mtmd_input_chunk_get_tokens_text(chunk, &count);
-        s.prompt_tokens.insert(s.prompt_tokens.end(), tokens, tokens + count);
-      } else if (type == MTMD_INPUT_CHUNK_TYPE_IMAGE) {
-        const int token_start = static_cast<int>(s.prompt_tokens.size());
-        const int token_count = static_cast<int>(
-            mtmd_input_chunk_get_n_tokens(chunk));
-        auto* copy = mtmd_input_chunk_copy(chunk);
-        if (copy == nullptr || token_count <= 0) {
-          if (copy) mtmd_input_chunk_free(copy);
-          return Result<PredictSetup>(std::unexpect,
-              make_error(ErrorCode::Internal,
-                         "multimodal prompt produced an invalid image chunk"));
-        }
-        s.prompt_tokens.insert(
-            s.prompt_tokens.end(), token_count, LLAMA_TOKEN_NULL);
-        s.media_chunks.push_back({
-            token_start,
-            token_count,
-            static_cast<int>(mtmd_input_chunk_get_n_pos(chunk)),
-            std::shared_ptr<mtmd_input_chunk>(
-                copy, [](mtmd_input_chunk* value) {
-                  mtmd_input_chunk_free(value);
-                }),
-        });
-      } else {
-        return Result<PredictSetup>(std::unexpect,
-            make_error(ErrorCode::InvalidArgument,
-                       "only image media is supported for chat inference"));
-      }
-    }
-    n_tokens = static_cast<int>(s.prompt_tokens.size());
-    if (s.prompt_tokens.empty() ||
-        s.prompt_tokens.back() == LLAMA_TOKEN_NULL) {
-      return Result<PredictSetup>(std::unexpect,
-          make_error(ErrorCode::Internal,
-                     "multimodal chat template did not produce a text generation suffix"));
-    }
-    s.checkpoint_capture_pos = 0;
-  } else {
-    s.prompt_tokens.resize(prompt.size() + 16);
-    n_tokens = llama_tokenize(
-        vocab_, prompt.data(), static_cast<int>(prompt.size()),
-        s.prompt_tokens.data(), static_cast<int>(s.prompt_tokens.size()),
-        add_bos, true);
-    if (n_tokens < 0) {
-      s.prompt_tokens.resize(static_cast<std::size_t>(-n_tokens));
-      n_tokens = llama_tokenize(
-          vocab_, prompt.data(), static_cast<int>(prompt.size()),
-          s.prompt_tokens.data(), static_cast<int>(s.prompt_tokens.size()),
-          add_bos, true);
-      if (n_tokens < 0) {
-        return Result<PredictSetup>(std::unexpect,
-            make_error(ErrorCode::Internal, "tokenization failed"));
-      }
-    }
-    s.prompt_tokens.resize(static_cast<std::size_t>(n_tokens));
-    s.prompt_position_count = n_tokens;
-    s.checkpoint_capture_pos = n_tokens;
-  }
-  const auto& generation_prompt = s.parser_params.generation_prompt;
-  if (!has_media && !generation_prompt.empty() &&
-      prompt.size() >= generation_prompt.size() &&
-      prompt.compare(
-          prompt.size() - generation_prompt.size(),
-          generation_prompt.size(),
-          generation_prompt) == 0) {
-    const std::string stable_prefix =
-        prompt.substr(0, prompt.size() - generation_prompt.size());
-    std::vector<llama_token> stable_tokens(stable_prefix.size() + 16);
-    int stable_count = llama_tokenize(
-        vocab_, stable_prefix.data(), static_cast<int>(stable_prefix.size()),
-        stable_tokens.data(), static_cast<int>(stable_tokens.size()),
-        add_bos, true);
-    if (stable_count < 0) {
-      stable_tokens.resize(static_cast<std::size_t>(-stable_count));
-      stable_count = llama_tokenize(
-          vocab_, stable_prefix.data(), static_cast<int>(stable_prefix.size()),
-          stable_tokens.data(), static_cast<int>(stable_tokens.size()),
-          add_bos, true);
-    }
-    if (stable_count >= 0) {
-      stable_tokens.resize(static_cast<std::size_t>(stable_count));
-      const int capture_pos = detail::recurrent_checkpoint_capture_pos(
-          s.prompt_tokens, stable_tokens);
-      if (capture_pos > 0) {
-        s.checkpoint_capture_pos = capture_pos;
-      }
-    }
-  }
-
-  // Per-slot context window = n_ctx_seq (total context / n_slots as set during load)
-  s.n_ctx_seq = static_cast<int>(llama_n_ctx_seq(shared_ctx_));
-
-  const int prompt_context = s.prompt_position_count > 0
-      ? s.prompt_position_count
-      : n_tokens;
-  if (prompt_context >= s.n_ctx_seq) {
-    if (!cfg_.truncate_prompt || has_media)
-      return Result<PredictSetup>(std::unexpect,
-          make_error(ErrorCode::ContextLengthExceeded,
-                     "This model's maximum context length is " + std::to_string(s.n_ctx_seq) +
-                     " tokens. However, your messages resulted in " + std::to_string(prompt_context) +
-                     " tokens. Please reduce the length of the messages."));
-    maybe_truncate_prompt(s.prompt_tokens, s.n_ctx_seq, req.max_output_tokens, info_.name);
-    n_tokens = static_cast<int>(s.prompt_tokens.size());
-    s.prompt_position_count = n_tokens;
-  }
-
-  const int ctx_budget = std::max(
-      1, s.n_ctx_seq - s.prompt_position_count - 1);
-  s.max_tokens = req.max_output_tokens > 0
-      ? std::min(req.max_output_tokens, ctx_budget) : ctx_budget;
-
-  // Snapshot per-slot KV state under the mutex (scheduler may touch these after submit)
-  {
-    std::lock_guard lk(mtx_);
-    s.last_prompt_tokens   = slots_[slot_id].last_prompt_tokens;
-    s.recurrent_checkpoint = slots_[slot_id].recurrent_checkpoint;
-    s.recurrent_draft_checkpoint = slots_[slot_id].recurrent_draft_checkpoint;
-    s.recurrent_mtp_checkpoint = slots_[slot_id].recurrent_mtp_checkpoint;
-    s.checkpoint_pos       = slots_[slot_id].checkpoint_pos;
-    s.mtp_cache_synced     = slots_[slot_id].mtp_cache_synced;
-  }
-
-  common_sampler* smp = common_sampler_init(model_, s.sampling_params);
-  if (smp == nullptr)
-    return Result<PredictSetup>(std::unexpect,
-        make_error(ErrorCode::Internal, "common_sampler_init returned null"));
-  s.smp = smp;
-
-  return Result<PredictSetup>(std::move(s));
-}
-
-// Drain task.out_queue until the done event, calling on_token for each token.
-// Returns error if the scheduler reported one.
-Result<void> LlamaCppModel::drain_task(SlotTask& task, const OnToken& on_token) {
-  while (true) {
-    TokenEvent ev;
-    {
-      std::unique_lock lk(task.out_mtx);
-      task.out_cv.wait(lk, [&task] { return !task.out_queue.empty(); });
-      ev = std::move(task.out_queue.front());
-      task.out_queue.pop();
-    }
-    if (ev.is_error)
-      return Result<void>(std::unexpect, make_error(ErrorCode::Internal, ev.error_msg));
-    if (ev.is_done)
-      return Result<void>{};
-    if (!on_token(ev.id) && !task.caller_stop.load())
-      task.caller_cancel.store(true);
-  }
-}
-
-Result<InferenceResult> LlamaCppModel::predict(int slot_id, const InferenceRequest& req) {
-  if (slot_id < 0 || slot_id >= static_cast<int>(slots_.size())) {
-    return Result<InferenceResult>(std::unexpect,
-        make_error(ErrorCode::InvalidArgument, "slot_id out of range"));
-  }
-  {
-    std::lock_guard lk(mtx_);
-    if (!loaded_.load())
-      return Result<InferenceResult>(std::unexpect, make_error(ErrorCode::Internal, "model not loaded"));
-    if (!slots_[slot_id].busy)
-      return Result<InferenceResult>(std::unexpect, make_error(ErrorCode::InvalidArgument, "slot not acquired"));
-  }
-
-  auto setup_res = prepare_inference(slot_id, req);
-  if (!setup_res.has_value())
-    return Result<InferenceResult>(std::unexpect, setup_res.error());
-  auto& setup = *setup_res;
-
-  const auto start = std::chrono::steady_clock::now();
-  std::string generated;
-  generated.reserve(4096);
-  std::vector<llama_token> decoded_ids;
-  bool string_stopped = false;
-
-  SlotTask task;
-  task.slot_id             = slot_id;
-  task.prompt_tokens       = setup.prompt_tokens;
-  task.media_chunks        = std::move(setup.media_chunks);
-  task.prompt_position_count = setup.prompt_position_count;
-  task.last_prompt_tokens  = setup.last_prompt_tokens;
-  task.recurrent_checkpoint = std::move(setup.recurrent_checkpoint);
-  task.recurrent_draft_checkpoint = std::move(setup.recurrent_draft_checkpoint);
-  task.recurrent_mtp_checkpoint = std::move(setup.recurrent_mtp_checkpoint);
-  task.checkpoint_pos       = setup.checkpoint_pos;
-  task.checkpoint_capture_pos = setup.checkpoint_capture_pos;
-  task.mtp_cache_synced    = setup.mtp_cache_synced;
-  task.sampler             = setup.smp;   // scheduler takes ownership
-  task.max_tokens          = setup.max_tokens;
-  task.stop_tokens         = setup.stop_tokens;
-
-  scheduler_->submit(&task);
-
-  auto drain_res = drain_task(task, [&](llama_token id) -> bool {
-    if (string_stopped) return false; // already stopping
-    generated.append(token_to_piece(vocab_, id));
-    for (const auto& stop : setup.stop_strings) {
-      if (!stop.empty() && generated.size() >= stop.size() &&
-          generated.compare(generated.size() - stop.size(), stop.size(), stop) == 0) {
-        generated.resize(generated.size() - stop.size());
-        string_stopped = true;
-        task.caller_stop.store(true);
-        return false; // do NOT push this token — it's part of the stop string
-      }
-    }
-    decoded_ids.push_back(id);
-    return true;
-  });
-  if (!drain_res.has_value())
-    return Result<InferenceResult>(std::unexpect, drain_res.error());
-
-  const auto end = std::chrono::steady_clock::now();
-
-  // Update per-slot KV state
-  {
-    std::lock_guard lk(mtx_);
-    auto& slot = slots_[slot_id];
-    slot.last_prompt_tokens.assign(setup.prompt_tokens.begin(), setup.prompt_tokens.end());
-    slot.last_prompt_tokens.insert(slot.last_prompt_tokens.end(),
-                                   decoded_ids.begin(), decoded_ids.end());
-    slot.recurrent_checkpoint = std::move(task.out_recurrent_checkpoint);
-    slot.recurrent_draft_checkpoint = std::move(task.out_recurrent_draft_checkpoint);
-    slot.recurrent_mtp_checkpoint = std::move(task.out_recurrent_mtp_checkpoint);
-    slot.checkpoint_pos       = task.out_checkpoint_pos;
-    slot.mtp_cache_synced     = task.out_mtp_cache_synced;
-  }
-
-  log_slot_release(info_.name,
-                   static_cast<int>(setup.prompt_tokens.size()),
-                   false,
-                   static_cast<int>(decoded_ids.size()),
-                   setup.n_ctx_seq);
-
-  InferenceResult out;
-  out.prompt_tokens         = static_cast<int>(setup.prompt_tokens.size());
-  out.cached_prompt_tokens  = task.out_cached_prompt_tokens;
-  out.completion_tokens     = static_cast<int>(decoded_ids.size());
-  out.duration_ms = std::chrono::duration<float, std::milli>(end - start).count();
-  out.prompt_duration_ms = task.out_prompt_duration_ms;
-  out.first_token_duration_ms = task.out_first_token_duration_ms;
-  out.generation_duration_ms = task.out_generation_duration_ms > 0.0f
-      ? task.out_generation_duration_ms
-      : out.duration_ms;
-  out.tokens_per_second = detail::generation_tokens_per_second(
-      out.completion_tokens, out.generation_duration_ms);
-  out.mtp_drafted_tokens = task.n_drafted;
-  out.mtp_accepted_tokens = task.n_draft_accepted;
-  if (out.completion_tokens >= setup.max_tokens)
-    out.finish_reason = "length";
-
-  try {
-    auto msg = parse_final_message_with_ids(generated, setup.parser_params);
-    apply_parsed_message(out, msg);
-    if (setup.parser_params.parse_tool_calls && out.tool_calls.empty())
-      apply_fallback_tool_calls(out, generated);
-  } catch (const std::exception& e) {
-    LOG_ERROR("chat_parse_failed", "model={} error={}", info_.name, e.what());
-    out.text = std::move(generated);
-    if (setup.parser_params.parse_tool_calls)
-      apply_fallback_tool_calls(out, out.text);
-  }
-  out.reasoning_tokens = count_output_tokens(vocab_, out.reasoning_text);
-  return Result<InferenceResult>(std::move(out));
-}
-
-Result<InferenceResult> LlamaCppModel::predict_stream(
-    int slot_id, const InferenceRequest& req, const TokenCallback& callback,
-    const std::atomic<bool>* cancel) {
-  if (slot_id < 0 || slot_id >= static_cast<int>(slots_.size())) {
-    return Result<InferenceResult>(std::unexpect,
-        make_error(ErrorCode::InvalidArgument, "slot_id out of range"));
-  }
-  {
-    std::lock_guard lk(mtx_);
-    if (!loaded_.load())
-      return Result<InferenceResult>(std::unexpect, make_error(ErrorCode::Internal, "model not loaded"));
-    if (!slots_[slot_id].busy)
-      return Result<InferenceResult>(std::unexpect, make_error(ErrorCode::InvalidArgument, "slot not acquired"));
-  }
-
-  auto setup_res = prepare_inference(slot_id, req);
-  if (!setup_res.has_value())
-    return Result<InferenceResult>(std::unexpect, setup_res.error());
-  auto& setup = *setup_res;
-
-  const auto start = std::chrono::steady_clock::now();
-  std::string generated;
-  generated.reserve(4096);
-  std::vector<llama_token> decoded_ids;
-  StreamingChatParserState parser_state(setup.parser_params);
-  detail::StreamingToolCallState tool_call_stream_state;
-  bool callback_aborted = false;
-  bool string_stopped = false;
-
-  SlotTask task;
-  task.slot_id            = slot_id;
-  task.prompt_tokens      = setup.prompt_tokens;
-  task.media_chunks       = std::move(setup.media_chunks);
-  task.prompt_position_count = setup.prompt_position_count;
-  task.last_prompt_tokens = setup.last_prompt_tokens;
-  task.recurrent_checkpoint = std::move(setup.recurrent_checkpoint);
-  task.recurrent_draft_checkpoint = std::move(setup.recurrent_draft_checkpoint);
-  task.recurrent_mtp_checkpoint = std::move(setup.recurrent_mtp_checkpoint);
-  task.checkpoint_pos      = setup.checkpoint_pos;
-  task.checkpoint_capture_pos = setup.checkpoint_capture_pos;
-  task.mtp_cache_synced    = setup.mtp_cache_synced;
-  task.sampler            = setup.smp;
-  task.max_tokens         = setup.max_tokens;
-  task.stop_tokens        = setup.stop_tokens;
-  task.ext_cancel         = cancel;
-
-  scheduler_->submit(&task);
-
-  auto drain_res = drain_task(task, [&](llama_token id) -> bool {
-    if (callback_aborted || string_stopped) return false;
-
-    std::string piece = token_to_piece(vocab_, id);
-    if (!piece.empty()) {
-      generated.append(piece);
-
-      for (const auto& stop : setup.stop_strings) {
-        if (!stop.empty() && generated.size() >= stop.size() &&
-            generated.compare(generated.size() - stop.size(), stop.size(), stop) == 0) {
-          generated.resize(generated.size() - stop.size());
-          string_stopped = true;
-          task.caller_stop.store(true);
-          return false; // do NOT push this token — it's part of the stop string
-        }
-      }
-
-      decoded_ids.push_back(id);
-      std::vector<common_chat_msg_diff> diffs;
-      try {
-        diffs = parser_state.update(piece, /*is_partial=*/true,
-                                    setup.parser_params.parse_tool_calls);
-      } catch (...) {}
-
-      for (const auto& diff : diffs) {
-        auto delta = to_delta(diff);
-        if (delta.content.empty() && delta.reasoning_text.empty() && delta.tool_calls.empty())
-          continue;
-        if (!callback(delta)) {
-          callback_aborted = true;
-          return false;
-        }
-        tool_call_stream_state.observe(delta);
-      }
-    }
-    return true;
-  });
-  if (!drain_res.has_value())
-    return Result<InferenceResult>(std::unexpect, drain_res.error());
-
-  const auto end = std::chrono::steady_clock::now();
-
-  // Update per-slot KV state
-  {
-    std::lock_guard lk(mtx_);
-    auto& slot = slots_[slot_id];
-    slot.last_prompt_tokens.assign(setup.prompt_tokens.begin(), setup.prompt_tokens.end());
-    slot.last_prompt_tokens.insert(slot.last_prompt_tokens.end(),
-                                   decoded_ids.begin(), decoded_ids.end());
-    slot.recurrent_checkpoint = std::move(task.out_recurrent_checkpoint);
-    slot.recurrent_draft_checkpoint = std::move(task.out_recurrent_draft_checkpoint);
-    slot.recurrent_mtp_checkpoint = std::move(task.out_recurrent_mtp_checkpoint);
-    slot.checkpoint_pos       = task.out_checkpoint_pos;
-    slot.mtp_cache_synced     = task.out_mtp_cache_synced;
-  }
-
-  log_slot_release(info_.name,
-                   static_cast<int>(setup.prompt_tokens.size()),
-                   false,
-                   static_cast<int>(decoded_ids.size()),
-                   setup.n_ctx_seq);
-
-  InferenceResult out;
-  out.prompt_tokens        = static_cast<int>(setup.prompt_tokens.size());
-  out.cached_prompt_tokens = task.out_cached_prompt_tokens;
-  out.completion_tokens    = static_cast<int>(decoded_ids.size());
-  out.duration_ms = std::chrono::duration<float, std::milli>(end - start).count();
-  out.prompt_duration_ms = task.out_prompt_duration_ms;
-  out.first_token_duration_ms = task.out_first_token_duration_ms;
-  out.generation_duration_ms = task.out_generation_duration_ms > 0.0f
-      ? task.out_generation_duration_ms
-      : out.duration_ms;
-  out.tokens_per_second = detail::generation_tokens_per_second(
-      out.completion_tokens, out.generation_duration_ms);
-  out.mtp_drafted_tokens = task.n_drafted;
-  out.mtp_accepted_tokens = task.n_draft_accepted;
-  if (out.completion_tokens >= setup.max_tokens)
-    out.finish_reason = "length";
-
-  if (callback_aborted) return Result<InferenceResult>(std::move(out));
-
-  bool fallback_tool_calls_used = false;
-  try {
-    auto final_diffs = parser_state.update("", /*is_partial=*/false,
-                                            setup.parser_params.parse_tool_calls);
-    for (const auto& diff : final_diffs) {
-      auto delta = to_delta(diff);
-      if (delta.content.empty() && delta.reasoning_text.empty() && delta.tool_calls.empty())
-        continue;
-      if (!callback(delta)) return Result<InferenceResult>(std::move(out));
-      tool_call_stream_state.observe(delta);
-    }
-    auto msg = parse_final_message_with_ids(generated, setup.parser_params);
-    apply_parsed_message(out, msg);
-    if (setup.parser_params.parse_tool_calls && out.tool_calls.empty()) {
-      apply_fallback_tool_calls(out, generated);
-      fallback_tool_calls_used = !out.tool_calls.empty();
-    }
-  } catch (const std::exception& e) {
-    LOG_ERROR("chat_parse_failed", "model={} error={}", info_.name, e.what());
-    out.text = std::move(generated);
-    if (setup.parser_params.parse_tool_calls) {
-      apply_fallback_tool_calls(out, out.text);
-      fallback_tool_calls_used = !out.tool_calls.empty();
-    }
-  }
-
-  if (fallback_tool_calls_used && tool_call_stream_state.should_emit_fallback()) {
-    for (std::size_t i = 0; i < out.tool_calls.size(); ++i) {
-      const auto& tc = out.tool_calls[i];
-      InferenceDelta delta;
-      ToolCallDelta tcd;
-      tcd.index = i; tcd.id = tc.id; tcd.type = "function";
-      tcd.function_name = tc.function_name;
-      tcd.function_arguments = tc.function_arguments;
-      delta.tool_calls.push_back(std::move(tcd));
-      if (!callback(delta)) break;
-    }
-  }
-  out.reasoning_tokens = count_output_tokens(vocab_, out.reasoning_text);
-  return Result<InferenceResult>(std::move(out));
-}
+#include "llama_cpp_prompt_cache.ipp"
+#include "llama_cpp_decoding.ipp"
 
 }
